% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apportion_lr}
\alias{apportion_lr}
\title{Apportion Least-Remainder Method}
\usage{
apportion_lr(vector.length, values, proportions, shuffled = FALSE)
}
\arguments{
\item{vector.length}{Length for the output vector.}

\item{values}{Values for the output vector.}

\item{proportions}{Proportion distribution with one number for each value. This
must sum to 1.}

\item{shuffled}{If `TRUE`, randomly shuffle the order of the vector.}
}
\description{
Apportions a vector of values given a specified frequency
             distribution of those values such that the length of the output
             is robust to rounding and other instabilities.
}
