% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.mod.vital.R
\name{arrivals.icm}
\alias{arrivals.icm}
\title{Arrivals: icm Module}
\usage{
arrivals.icm(dat, at)
}
\arguments{
\item{dat}{Master data list object.}

\item{at}{Current time step.}
}
\description{
This function simulates arrival for use in `\link{icm}`
             simulations.
}
\seealso{
`\link{icm}`
}
\keyword{internal}
