% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.vital.R
\name{arrivals.net.grp}
\alias{arrivals.net.grp}
\title{Arrivals: netsim Module}
\usage{
arrivals.net.grp(dat, at)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{at}{Current time step.}
}
\description{
This function simulates new arrivals into the network
  for use in \code{\link{netsim}} simulations.
}
\seealso{
\code{\link{netsim}}
}
\keyword{internal}
\keyword{netMod}
