% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.epi.data.frame}
\alias{as.epi.data.frame}
\title{Validate and Convert to epi.data.frame}
\usage{
as.epi.data.frame(df)
}
\description{
This methods ensures that the \code{data.frame} is correctly formatted as an
\code{epi.data.frame}
}
