% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dendo.R
\name{as.network.transmat}
\alias{as.network.transmat}
\title{converting transmat infection tree into a network object}
\usage{
\method{as.network}{transmat}(tm)
}
\description{
converting transmat infection tree into a network object
}

