% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dendo.R
\name{as.phylo.tPath}
\alias{as.phylo.tPath}
\title{functions for converting to and from phylo class}
\usage{
\method{as.phylo}{tPath}(path)
}
\description{
functions for converting to and from phylo class
}

