% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dendo.R
\name{as.phylo.transmat}
\alias{as.phylo.transmat}
\title{convert transmat infection tree into a phylo tree object}
\usage{
\method{as.phylo}{transmat}(tm)
}
\description{
convert transmat infection tree into a phylo tree object
}

