% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.infection.R
\name{as_tibble_edgelist}
\alias{as_tibble_edgelist}
\title{Convert an Edgelist into a Tibble}
\usage{
as_tibble_edgelist(el)
}
\arguments{
\item{el}{An edgelist in matrix or data frame form.}
}
\value{
The edgelist in tibble form with two columns named \code{head} and \code{tail}.
}
\description{
Convert an Edgelist into a Tibble
}
