% Generated by roxygen2 (4.0.2): do not edit by hand
\name{births.icm}
\alias{births.icm}
\title{Births: icm Module}
\usage{
births.icm(dat, at)
}
\arguments{
\item{dat}{master data list object.}

\item{at}{current time step.}
}
\description{
This function simulates birth for use in \code{\link{icm}}
             simulations.
}
\seealso{
\code{\link{icm}}
}
\keyword{internal}

