% Generated by roxygen2 (4.0.2): do not edit by hand
\name{births.net}
\alias{births.net}
\title{Births: netsim Module}
\usage{
births.net(dat, at)
}
\arguments{
\item{dat}{a list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{at}{current time step.}
}
\description{
This function simulates new births into the network
  for use in \code{\link{netsim}} simulations.
}
\seealso{
\code{\link{netsim}}
}
\keyword{internal}
\keyword{netMod}

