\name{births}
\alias{births}
\title{Births: epiNet.simTrans Module}
\usage{
births(all, at)
}
\arguments{
  \item{all}{a list object containing a
  \code{networkDynamic} object and other initialization
  information passed from \code{\link{epiNet.simTrans}}.}

  \item{at}{time step for simulation.}
}
\description{
This function simulates new births into the network for use
in \code{\link{epiNet.simTrans}} simulations.
}
\seealso{
\code{\link{epiNet.simTrans}}
}
\keyword{epiNetModule}
\keyword{internal}

