% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.end.horizon.R
\name{check_end_horizon_control}
\alias{check_end_horizon_control}
\title{Check the format of the end.horizon control}
\usage{
check_end_horizon_control(dat)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link[=netsim]{netsim()}}.}
}
\value{
TRUE invisibly on succes, errors otherwise
}
\description{
Check the format of the end.horizon control
}
\keyword{internal}
