% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{color_tea}
\alias{color_tea}
\title{Creates a TEA Variable for Infection Status for `ndtv` Animations}
\usage{
color_tea(
  nd,
  old.var = "testatus",
  old.sus = "s",
  old.inf = "i",
  old.rec = "r",
  new.var = "ndtvcol",
  new.sus,
  new.inf,
  new.rec,
  verbose = TRUE
)
}
\arguments{
\item{nd}{An object of class `networkDynamic`.}

\item{old.var}{Old TEA variable name.}

\item{old.sus}{Status value for susceptible in old TEA variable.}

\item{old.inf}{Status value for infected in old TEA variable.}

\item{old.rec}{Status value for recovered in old TEA variable.}

\item{new.var}{New TEA variable name to be stored in `networkDynamic`
object.}

\item{new.sus}{Status value for susceptible in new TEA variable.}

\item{new.inf}{Status value for infected in new TEA variable.}

\item{new.rec}{Status value for recovered in new TEA variable.}

\item{verbose}{Print progress to console.}
}
\description{
Creates a new color-named temporally-extended attribute (TEA)
             variable in a `networkDynamic` object containing a disease
             status TEA in numeric format.
}
\details{
The `ndtv` package (\url{https://cran.r-project.org/package=ndtv}) produces
animated visuals for dynamic networks with evolving edge structures and nodal
attributes. Nodal attribute dynamics in `ndtv` movies require a temporally
extended attribute (TEA) containing a standard R color for each node at each
time step. By default, the `EpiModel` package uses TEAs to store disease
status history in network model simulations run in `\link{netsim}`. But,
that status TEA is in numeric format (0, 1, 2). The `color_tea` function
transforms those numeric values of that disease status TEA into a TEA with
color values in order to visualize status changes in `ndtv`.

The convention in `\link{plot.netsim}` is to color the susceptible
nodes as blue, infected nodes as red, and recovered nodes as green. Alternate
colors may be specified using the `new.sus`, `new.inf`, and
`new.rec` parameters, respectively.

Using the `color_tea` function with a `netsim` object requires that
TEAs for disease status be used and that the `networkDynamic` object be
saved in the output: `tergmListe` must be  set to `FALSE` in
`\link{control.net}`.
}
\seealso{
`\link{netsim}` and the `ndtv` package documentation.
}
\keyword{colorUtils}
