% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.updater.R
\name{common_updater}
\alias{common_updater}
\title{Update either the "param" or "control" list}
\usage{
common_updater(dat, type)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}

\item{type}{either "param" or "control"}
}
\value{
The updated \code{dat} Master list object.
}
\description{
Update either the "param" or "control" list
}
\keyword{internal}
