% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.inputs.R
\name{control.icm}
\alias{control.icm}
\title{Control Settings for Stochastic Individual Contact Models}
\usage{
control.icm(type, nsteps, nsims = 1, verbose = FALSE, verbose.int = 0)
}
\arguments{
\item{type}{Disease type to be modeled, with the choice of \code{"SI"} for
Susceptible-Infected diseases, \code{"SIR"} for
Susceptible-Infected-Recovered diseases, and \code{"SIS"} for
Susceptible-Infected-Susceptible diseases.}

\item{nsteps}{Number of time steps to solve the model over. This must be a
positive integer.}

\item{nsims}{Number of simulations to run.}

\item{verbose}{If \code{TRUE}, print model progress to the console.}

\item{verbose.int}{Time step interval for printing progress to console, where
0 (the default) prints completion status of entire simulation and
positive integer \code{x} prints progress after every \code{x} time
steps.}
}
\value{
An \code{EpiModel} object of class \code{control.icm}.
}
\description{
Sets the controls for stochastic individual contact models
simulated with \code{\link[=icm]{icm()}}.
}
\details{
\code{control.icm} sets the required control settings for any stochastic
individual contact model solved with the \code{\link[=icm]{icm()}} function. ICM simulations
use the built-in SI, SIR, and SIS disease types only. The \code{type} argument
is required and has no default. For custom or extension epidemic models, use
the network model class via \code{\link[=control.net]{control.net()}} instead.
}
\seealso{
Use \code{\link[=param.icm]{param.icm()}} to specify model parameters and
\code{\link[=init.icm]{init.icm()}} to specify the initial conditions. Run the
parameterized model with \code{\link[=icm]{icm()}}.
}
\keyword{parameterization}
