% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{copy_datattr_to_nwattr}
\alias{copy_datattr_to_nwattr}
\title{Copies Vertex Attributes from the dat List to the Network Object}
\usage{
copy_datattr_to_nwattr(dat)
}
\arguments{
\item{dat}{Master data object passed through `netsim` simulations.}
}
\description{
Copies the vertex attributes stored on the master attr list on
             dat to the network object on dat.
}
\seealso{
`\link{get_formula_term_attr}`, `\link{get_attr_prop}`,
         `\link{auto_update_attr}`, and `\link{copy_nwattr_to_datattr}`.
}
\keyword{internal}
\keyword{netUtils}
