% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{copy_nwattr_to_datattr}
\alias{copy_nwattr_to_datattr}
\title{Copies Vertex Attributes From Network to dat List}
\usage{
copy_nwattr_to_datattr(dat)
}
\arguments{
\item{dat}{Master data object passed through `netsim` simulations.}
}
\description{
Copies the vertex attributes stored on the network object to the
             master attr list in the dat data object.
}
\seealso{
`\link{get_formula_term_attr}`, `\link{get_attr_prop}`,
         `\link{auto_update_attr}`, and `\link{copy_datattr_to_nwattr}`.
}
\keyword{internal}
\keyword{netUtils}
