% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{copy_toall_attr.bip}
\alias{copy_toall_attr.bip}
\title{Copies Vertex Attributes in Formation Formula to attr List}
\usage{
copy_toall_attr.bip(dat, at, fterms)
}
\arguments{
\item{dat}{Master data object passed through \code{netsim.bip} simulations.}

\item{at}{Current time step.}

\item{fterms}{Vector of attributes used in formation formula, usually as
output of \code{\link{get_formula_term_attr.bip}}.}
}
\description{
Copies the vertex attributes stored on the network object to the
             master attr list in the dat data object.
}
\seealso{
\code{\link{get_formula_term_attr.bip}}, \code{\link{get_attr_prop.bip}},
         \code{\link{update_nwattr.bip}}.
}
\keyword{internal}
\keyword{netUtils}
