% Generated by roxygen2 (4.0.2): do not edit by hand
\name{copy_toall_attr}
\alias{copy_toall_attr}
\title{Copies Vertex Attributes in Formation Formula to attr List}
\usage{
copy_toall_attr(dat, at, t)
}
\arguments{
\item{dat}{master data object passed through \code{netsim} simulations.}

\item{at}{current time step.}

\item{t}{vector of attributes used in formation formula, usually as output of
       \code{\link{get_formula_terms}}.}
}
\description{
Copies the vertex attributes stored on the network object to the
             master attr list in the dat data object.
}
\seealso{
\code{\link{get_formula_terms}}, \code{\link{get_attr_prop}},
         \code{\link{update_nwattr}}.
}
\keyword{internal}
\keyword{netUtils}

