% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.helpers.R
\name{create_dat_object}
\alias{create_dat_object}
\title{Create a Minimal Master List Object of Network Model}
\usage{
create_dat_object(param = list(), init = list(), control = list())
}
\arguments{
\item{param}{An \code{EpiModel} object of class \code{\link{param.net}}.}

\item{init}{An \code{EpiModel} object of class \code{\link{init.net}}.}

\item{control}{An \code{EpiModel} object of class \code{\link{control.net}}.}
}
\value{
A \code{dat} Master list object
}
\description{
This helper function populate a \code{dat} Master List object
             with the minimal required elements. All parameters are optional.
             When none is given the resulting object is only a shell list
             with the different named elements defined as empty lists.
}
