% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosscheck.R
\name{crosscheck.dcm.bip}
\alias{crosscheck.dcm.bip}
\title{Cross Checking of Inputs for Deterministic Bipartite Compartmental Models}
\usage{
crosscheck.dcm.bip(param, init, control)
}
\arguments{
\item{param}{An \code{EpiModel} object of class \code{\link{param.dcm.bip}}.}

\item{init}{An \code{EpiModel} object of class \code{\link{init.dcm.bip}}.}

\item{control}{An \code{EpiModel} object of class \code{\link{control.dcm.bip}}.}
}
\value{
This function returns no objects.
}
\description{
This function checks that the three parameter lists from
             \code{\link{param.dcm.bip}}, \code{\link{init.dcm.bip}}, and
             \code{\link{control.dcm.bip}} are consistent.
}
\keyword{internal}
