% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm.inputs.R
\name{crosscheck.dcm}
\alias{crosscheck.dcm}
\title{Cross Checking of Inputs for Deterministic Compartmental Models}
\usage{
crosscheck.dcm(param, init, control)
}
\arguments{
\item{param}{An `EpiModel` object of class `\link{param.dcm}`.}

\item{init}{An `EpiModel` object of class `\link{init.dcm}`.}

\item{control}{An `EpiModel` object of class `\link{control.dcm}`.}
}
\value{
This function returns no objects.
}
\description{
This function checks that the three parameter lists from
             `\link{param.dcm}`, `\link{init.dcm}`, and
             `\link{control.dcm}` are consistent.
}
\keyword{internal}
