% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosscheck.R
\name{crosscheck.icm.bip}
\alias{crosscheck.icm.bip}
\title{Cross Checking of Inputs for Stochastic Bipartite Individual Contact Models}
\usage{
crosscheck.icm.bip(param, init, control)
}
\arguments{
\item{param}{An \code{EpiModel} object of class \code{\link{param.icm.bip}}.}

\item{init}{An \code{EpiModel} object of class \code{\link{init.icm.bip}}.}

\item{control}{An \code{EpiModel} object of class \code{\link{control.icm.bip}}.}
}
\value{
This function returns no objects.
}
\description{
This function checks that the three parameter lists from
             \code{\link{param.icm.bip}}, \code{\link{init.icm.bip}}, and
             \code{\link{control.icm.bip}} are consistent.
}
\keyword{internal}
