% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.inputs.R
\name{crosscheck.icm}
\alias{crosscheck.icm}
\title{Cross Checking of Inputs for Stochastic Individual Contact Models}
\usage{
crosscheck.icm(param, init, control)
}
\arguments{
\item{param}{An \code{EpiModel} object of class \code{\link[=param.icm]{param.icm()}}.}

\item{init}{An \code{EpiModel} object of class \code{\link[=init.icm]{init.icm()}}.}

\item{control}{An \code{EpiModel} object of class \code{\link[=control.icm]{control.icm()}}.}
}
\value{
This function returns no objects.
}
\description{
This function checks that the three parameter lists from
\code{\link[=param.icm]{param.icm()}}, \code{\link[=init.icm]{init.icm()}}, and
\code{\link[=control.icm]{control.icm()}} are consistent.
}
\keyword{internal}
