% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.inputs.R
\name{crosscheck.net}
\alias{crosscheck.net}
\title{Cross Checking of Inputs for Stochastic Network Models}
\usage{
crosscheck.net(x, param, init, control)
}
\arguments{
\item{x}{An `EpiModel` object of class `\link{netest}`.}

\item{param}{An `EpiModel` object of class `\link{param.net}`.}

\item{init}{An `EpiModel` object of class `\link{init.net}`.}

\item{control}{An `EpiModel` object of class `\link{control.net}`.}
}
\value{
This function returns no objects.
}
\description{
This function checks that the estimation object from
             `\link{netest}` and the three parameter lists from
             `\link{param.net}`, `\link{init.net}`, and
             `\link{control.net}` are consistent.
}
\keyword{internal}
