% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dcm.mods.R
\name{dcm.mods}
\alias{dcm.mods}
\alias{mod_SIR_1g_cl}
\alias{mod_SIR_1g_op}
\alias{mod_SIR_2g_cl}
\alias{mod_SIR_2g_op}
\alias{mod_SIS_1g_cl}
\alias{mod_SIS_1g_op}
\alias{mod_SIS_2g_cl}
\alias{mod_SIS_2g_op}
\alias{mod_SI_1g_cl}
\alias{mod_SI_1g_op}
\alias{mod_SI_2g_cl}
\alias{mod_SI_2g_op}
\title{Deterministic Compartmental Model Functions}
\usage{
mod_SI_1g_cl(t, t0, parms)

mod_SI_1g_op(t, t0, parms)

mod_SI_2g_cl(t, t0, parms)

mod_SI_2g_op(t, t0, parms)

mod_SIR_1g_cl(t, t0, parms)

mod_SIR_1g_op(t, t0, parms)

mod_SIR_2g_cl(t, t0, parms)

mod_SIR_2g_op(t, t0, parms)

mod_SIS_1g_cl(t, t0, parms)

mod_SIS_1g_op(t, t0, parms)

mod_SIS_2g_cl(t, t0, parms)

mod_SIS_2g_op(t, t0, parms)
}
\arguments{
\item{t}{Time vector.}

\item{t0}{Initial conditions.}

\item{parms}{Model parameters.}
}
\description{
These functions parameterize the built-in deterministic
             compartmental models solved using the \code{\link{dcm}}
             function in \code{EpiModel}.
}
\details{
Details to add.
}
\keyword{internal}

