% Generated by roxygen2 (4.0.1): do not edit by hand
\name{deaths.icm}
\alias{deaths.icm}
\title{Deaths: icm Module}
\usage{
deaths.icm(all, at)
}
\arguments{
\item{all}{master data list object.}

\item{at}{current time step.}
}
\description{
This function simulates death for use in \code{\link{icm}}
             simulations.
}
\seealso{
\code{\link{icm}}
}
\keyword{internal}

