\name{deaths.inf}
\alias{deaths.inf}
\title{Deaths for Infecteds: epiNet.simTrans Module}
\usage{
deaths.inf(all, at)
}
\arguments{
  \item{all}{a list object containing a
  \code{networkDynamic} object and other initialization
  information passed from \code{\link{epiNet.simTrans}}.}

  \item{at}{time step for simulation.}
}
\description{
This function simulates deaths for infecteds for use in
\code{\link{epiNet.simTrans}} simulations.
}
\seealso{
\code{\link{epiNet.simTrans}}, \code{\link{deaths.sus}},
\code{\link{deaths.rec}}
}
\keyword{epiNetModule}
\keyword{internal}

