% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/net.mod.vital.R
\name{deaths.net}
\alias{deaths.net}
\title{Deaths: netsim Module}
\usage{
deaths.net(dat, at)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{at}{Current time step.}
}
\description{
This function simulates death for use in \link{netsim} simulations.
}
\seealso{
\code{\link{netsim}}
}
\keyword{internal}
\keyword{netMod}

