% Generated by roxygen2 (4.0.1): do not edit by hand
\name{deaths_rec.net}
\alias{deaths_rec.net}
\title{Deaths for Recovered: netsim Module}
\usage{
deaths_rec.net(all, at)
}
\arguments{
\item{all}{a list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{at}{current time step.}
}
\description{
This function simulates death for those in the susceptible state
  for use in \code{\link{netsim}} simulations.
}
\seealso{
\code{\link{netsim}}, \code{\link{deaths_sus.net}},
 \code{\link{deaths_inf.net}}
}
\keyword{internal}
\keyword{netMod}

