% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete_attr}
\alias{delete_attr}
\title{Delete Elements from Attribute List}
\usage{
delete_attr(dat, ids)
}
\arguments{
\item{dat}{Either the \code{netsim_dat} class object passed through
\code{netsim} simulations, or the main data object passed through
\code{icm} simulations.}

\item{ids}{ID numbers to delete from the list.}
}
\value{
The updated \code{netsim_dat} main list object.
}
\description{
Deletes elements from the main attribute list.
}
\keyword{internal}
