% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{delete_edges}
\alias{delete_edges}
\title{Remove Edges That Include Specified Vertices}
\usage{
delete_edges(el, vid)
}
\arguments{
\item{el}{A two-column matrix of current edges (edgelist).}

\item{vid}{A vector of vertex IDs whose edges are to be deleted from the
edgelist.}
}
\value{
Returns an updated edgelist object, with any edges including the specified
vertices removed.
}
\description{
Given a current two-column matrix of edges and a vector of
vertex IDs, this function removes any rows of the edgelist in
which the IDs are present.
}
