% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deleteAttr.bip}
\alias{deleteAttr.bip}
\title{Delete Elements from Attribute List}
\usage{
deleteAttr.bip(attrList, ids)
}
\arguments{
\item{attrList}{Attribute list.}

\item{ids}{ID numbers to delete from the list.}
}
\description{
Deletes elements from the master attribute list.
}
\keyword{internal}
