% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{depart_nodes}
\alias{depart_nodes}
\title{Depart Nodes from the netsim_dat Object}
\usage{
depart_nodes(dat, departures)
}
\arguments{
\item{dat}{the \code{netsim_dat} object}

\item{departures}{the vertex ids of nodes to depart}
}
\value{
the updated \code{netsim_dat} object with the nodes in
\code{departures} departed
}
\description{
Depart Nodes from the netsim_dat Object
}
\details{
If \code{tergmLite} is \code{FALSE}, the vertex ids
\code{departures} are deactivated (from the current timestep onward) in each
\code{networkDynamic} stored in \code{dat$run$nw}. If \code{tergmLite} is
\code{TRUE}, the vertex ids \code{departures} are deleted from \code{dat$run$el},
\code{dat$run$attr}, and \code{dat$run$net_attr}.
}
