% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.mod.vital.R
\name{departures.icm.bip}
\alias{departures.icm.bip}
\title{Departure: Bipartite icm Module}
\usage{
departures.icm.bip(dat, at)
}
\arguments{
\item{dat}{Main data object passed through \code{icm} simulations.}

\item{at}{Current time step.}
}
\value{
The updated \code{netsim_dat} main list object.
}
\description{
This function simulates departure for use in \code{\link{icm}}
simulations.
}
\seealso{
\code{\link{icm}}
}
\keyword{internal}
