% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.mod.vital.R
\name{departures.icm}
\alias{departures.icm}
\title{Departure: icm Module}
\usage{
departures.icm(dat, at)
}
\arguments{
\item{dat}{Master data list object.}

\item{at}{Current time step.}
}
\description{
This function simulates departure for use in `\link{icm}`
             simulations.
}
\seealso{
`\link{icm}`
}
\keyword{internal}
