% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.vital.R
\name{departures.net.bip}
\alias{departures.net.bip}
\title{Departures: netsim.bip Module}
\usage{
departures.net.bip(dat, at)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim.bip}}.}

\item{at}{Current time step.}
}
\description{
This function simulates departure for use in \link{netsim.bip} simulations.
}
\seealso{
\code{\link{netsim}}
}
\keyword{internal}
\keyword{netMod}
