% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.vital.R
\name{departures.net}
\alias{departures.net}
\title{Departures: netsim Module}
\usage{
departures.net(dat, at)
}
\arguments{
\item{dat}{Master list object containing a `networkDynamic` object and other
initialization information passed from `\link{netsim}`.}

\item{at}{Current time step.}
}
\description{
This function simulates departure for use in `\link{netsim}`` simulations.
}
\seealso{
`\link{netsim}`
}
\keyword{internal}
\keyword{netMod}
