% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{edgelist_censor.bip}
\alias{edgelist_censor.bip}
\title{Table of Edge Censoring}
\usage{
edgelist_censor.bip(el)
}
\arguments{
\item{el}{Timed edgelist with start and end times extracted from a
\code{networkDynamic.bip} object using the \code{as.data.frame.networkDynamic}
function.}
}
\description{
Outputs a table of the number and percent of edges that are
             left-censored, right-censored, both-censored, or uncensored for a
             \code{networkDynamic.bip} object.
}
\details{
Given a STERGM simulation over a specified number of time steps, the edges
within that simulation may be left-censored (started before the first step),
right-censored (continued after the last step), right and left-censored, or
uncensored. The amount of censoring will increase when the average edge
duration approaches the length of the simulation.
}
\examples{
# Initialize and parameterize bipartite network model
nw <- network.initialize.bip(n = 100, directed = FALSE)
formation <- ~edges
target.stats <- 50
coef.diss <- dissolution_coefs.bip(dissolution = ~offset(edges), duration = 20)

# Model estimation
est <- netest.bip(nw, formation, target.stats, coef.diss, verbose = FALSE)

# Simulate the network and extract a timed edgelist
dx <- netdx.bip(est, nsims = 1, nsteps = 100, keep.tedgelist = TRUE, verbose = FALSE)
el <- as.data.frame(dx)

# Calculate censoring
edgelist_censor.bip(el)

}
\keyword{netUtils}
