% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{edgelist_meanage}
\alias{edgelist_meanage}
\title{Mean Age of Partnerships over Time}
\usage{
edgelist_meanage(el, diss_term = NULL, diss_attr = NULL)
}
\arguments{
\item{el}{A timed edgelist with start and end times extracted from a
\code{networkDynamic} object using the
\code{as.data.frame.networkDynamic} function.}

\item{diss_term}{A string indicating the form of heterogeneity present 
in the dissolution model (options \code{nodematch} and \code{nodemix}),
or \code{NULL} for homogeneous (edges-only) dissolution model}

\item{attribute}{A vector containing values of the nodal attribute
associated with the heterogeneity in the dissolution model,
or \code{NULL} for homogeneous (edges-only) dissolution model}
}
\value{
A matrix containing the mean edge age at each timestep (rows),
with either one column (for homogeneous models, i.e. when 
\code{diss_term=NULL}) or one column per attribute value combination 
(for heterogeneous models).
}
\description{
Outputs a matrix of mean ages of edges at a series of timesteps.
}
\details{
This function calculates the mean partnership age at each time step over
a dynamic network simulation expressed in the form of a timed edgelist. 
Means may be calculated for all edges, or disaggregated by nodal attribute  
combinations.
}
\keyword{internal}
\keyword{netUtils}
