% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.simnet.R
\name{edges_correct.bip}
\alias{edges_correct.bip}
\title{Adjustment for the Edges Coefficient with Changing Network Size}
\usage{
edges_correct.bip(dat, at)
}
\arguments{
\item{dat}{Master object in \code{netsim.bip} simulations.}

\item{at}{Current time step.}
}
\description{
Adjusts the edges coefficient in a dynamic network model
             simulated in \code{\link{netsim.bip}} to preserve the mean
             degree of nodes in the network.
}
\keyword{internal}
