% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{geom_bands.bip}
\alias{geom_bands.bip}
\title{ggplot2 geom for Quantile Bands}
\usage{
geom_bands.bip(mapping, lower = 0.25, upper = 0.75, alpha = 0.25,
  ...)
}
\arguments{
\item{mapping}{standard aesthetic mapping \code{aes()} input for ggplot2.}

\item{lower}{Lower quantile for the time series.}

\item{upper}{Upper quantile for the time series.}

\item{alpha}{Transparency of the ribbon fill.}

\item{...}{Additional arguments passed to \code{stat_summary}.}
}
\description{
Plots quantile bands given a data.frame with stochastic model
             results from \code{icm.bip} or \code{netsim.bip}.
}
\details{
This is a wrapper around \code{ggplot::stat_summary} with a ribbon geom as
aesthetic output.
}
\examples{
param <- param.icm.bipinf.prob = 0.2, act.rate = 0.25) (Need Example)
init <- init.icm.bip(s.num = 500, i.num = 1)
control <- control.icm.bip(type = "SI", nsteps = 250, nsims = 5)
mod1 <- icm.bip(param, init, control)
df <- as.data.frame(mod1)
df.mean <- as.data.frame(mod1, out = "mean")

library(ggplot2)
ggplot() +
   geom_line(data = df, mapping = aes(time, i.num, group = sim), alpha = 0.25,
             lwd = 0.25, color = "firebrick") +
   geom_bands(data = df, mapping = aes(time, i.num),
              lower = 0.1, upper = 0.9, fill = "firebrick") +
   geom_line(data = df.mean, mapping = aes(time, i.num)) +
   theme_minimal()

}
\keyword{plot}
