% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_args}
\alias{get_args}
\title{Get Arguments from Parent Function}
\usage{
get_args(...)
}
\arguments{
\item{...}{dot arguments in function.}
}
\description{
Gets the arguments and values from the parent function environment
             and returns them in a list.
}
\keyword{internal}

