% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.dat.R
\name{get_attr}
\alias{get_attr}
\title{Get the nodale attributes of a Master list object in network models}
\usage{
get_attr(dat, attr_names = NULL, list_out = FALSE)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{attr_names}{the name of the attribute(s) to get (default = NULL)}

\item{list_out}{Should the output be a list containing the element(s)?
See **Value**. (default = FALSE)}
}
\value{
If `attr_names` is null, the full nodal attribute list. If `attr_names`
        is of length one and `list_out == FALSE` (default), the `attr_names`
        element of the nodal attribute list. If `list_out == TRUE`, a list
        containing the `attr_names` element(s)  of the nodale attribute list.
}
\description{
Helper function to access the nodale attributes list of the
             Master list object passed by modules during the course of a
             network simulation
}
\examples{

```
get_attr(dat)
get_attr(dat, "age")
get_attr(dat, "age", list_out = TRUE)
get_attr(dat, c("age", "status"), list_out = TRUE)
```

}
