% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_attr_history}
\alias{get_attr_history}
\title{Extract the Attributes History from Network Simulations}
\usage{
get_attr_history(sims)
}
\arguments{
\item{sims}{An \code{EpiModel} object of class \code{netsim}.}
}
\value{
A list of \code{data.frame}s, one for each "measure" recorded in the
simulation by the `record_attr_history` function.
}
\description{
Extract the Attributes History from Network Simulations
}
\examples{
\dontrun{

# With `sims` the result of a `netsim` call
get_attr_history(sims)

}

}
