% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reachable.R
\name{get_connected_subnet}
\alias{get_connected_subnet}
\title{Returns all the node connected directly or indirectly to a set of nodes}
\usage{
get_connected_subnet(adj_list, nodes)
}
\arguments{
\item{adj_list}{The network represented as an adjacency list}

\item{nodes}{A set of nodes}
}
\value{
A vector of nodes indexes that are connected together with the ones
provided in the \code{nodes} argument
}
\description{
Returns all the node connected directly or indirectly to a set of nodes
}
