% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.helpers.R
\name{get_current_timestep}
\alias{get_current_timestep}
\title{Returns the Current Timestep}
\usage{
get_current_timestep(dat)
}
\arguments{
\item{dat}{a Master list object of network models}
}
\value{
the current timestep
}
\description{
Returns the Current Timestep
}
