% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.dat.R
\name{get_dat_elt}
\alias{get_dat_elt}
\title{internal helper function to get elements of a Master list object}
\usage{
get_dat_elt(dat, elt, attrnames, list_out)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{elt}{the name of the sublist to access}

\item{attrnames}{the name(s) of the `elt` list elements to get}

\item{list_out}{Should the output be a list containing the element(s)?
See **Value**.}
}
\value{
If `attrname` is null, the full `elt` list.
  If `attrname` is of length one and `list_out == FALSE` (default),
  the `attrname` elements of the `elt` list.
  If `list_out == TRUE`, a list containing the `attrname` element(s)
  of the `elt` list.
}
\description{
internal helper function to get elements of a Master list object
}
\examples{

```
get_dat_elt(dat, "attr")
get_dat_elt(dat, "attr", "age")
get_dat_elt(dat, "attr", "age", list_out = TRUE)
get_dat_elt(dat, "attr", c("age", "status"), list_out = TRUE)
```

}
\keyword{Internal}
