% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.infection.R
\name{get_discordant_edgelist}
\alias{get_discordant_edgelist}
\title{Get Discordant Edgelist Based on Specified Status Variable}
\usage{
get_discordant_edgelist(
  dat,
  status.attr,
  head.status = "i",
  tail.status = "s",
  networks = NULL
)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{netsim}}.}

\item{status.attr}{The name of the status attribute of interest. This parameter is required.}

\item{head.status}{The value(s) of \code{status.attr} for which to look for the head of the edge.
Can be a single value or a vector. Default of \code{head.status = "i"}.}

\item{tail.status}{The value(s) of \code{status.attr} for which to look for the tail of the edge.
Can be a single value or a vector. Default of \code{tail.status = "s"}.}

\item{networks}{In case of models with multiple networks, the network(s) from which to pull
the current edgelist. Can be a single value or a vector. If \code{NULL} (the default),
then all networks will be included.}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{head}: Positional ID of the head node.
\item \code{tail}: Positional ID of the tail node.
\item \code{network}: The numerical index of the network on which the partnership is located.
\item \code{head_status}: Status of the head node.
\item \code{tail_status}: Status of the tail node.
}
}
\description{
This function returns a \code{data.frame} with a discordant
edgelist, defined as the set of edges for which the status attribute
of interest is infected for one partner and susceptible for the other.
}
\details{
This is a generalized version of the \code{discord_edgelist} function.
It creates an edgelist of current partnerships in which the status attribute
of interest (as specified by the parameter \code{status.attr}) of one partner matches
the value (or one of the values) of the \code{head.status} parameter while the
corresponding status attribute of the other partner matches  the value (or
one of the values) of \code{tail.status} parameter.
}
\seealso{
\code{\link{discord_edgelist}}
}
\keyword{internal}
\keyword{netMod}
