% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_cuml.R
\name{get_edgelist}
\alias{get_edgelist}
\title{Get an Edgelist From the Specified Network}
\usage{
get_edgelist(dat, network)
}
\arguments{
\item{dat}{Master list object of network models.}

\item{network}{Numerical index of the network from which the edgelist should
be extracted.}
}
\value{
An edgelist in matrix form with the two columns. Each column contains the
posit_ids (see \code{get_posit_ids})of the nodes in each edge.
}
\description{
This function outputs an edgelist from the specified network,
             selecting the method depending on the stored network type.
}
