% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelists.R
\name{get_edgelists_df}
\alias{get_edgelists_df}
\title{Get the Edgelist(s) from the Specified Network(s)}
\usage{
get_edgelists_df(dat, networks = NULL)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{netsim}}.}

\item{networks}{Numerical indexes of the networks to extract the partnerships
from. (May be > 1 for models with multiple overlapping
networks.) If \code{NULL}, extract from all networks.}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{head}: Positional ID of the head node.
\item \code{tail}: Positional ID of the tail node.
\item \code{network}: The numerical index of the network on which the edge is located.
}
}
\description{
Get the Edgelist(s) from the Specified Network(s)
}
