% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.dat.R
\name{get_epi}
\alias{get_epi}
\title{Get the Epidemic Outputs of a Master list object in network models}
\usage{
get_epi(dat, epi_names = NULL, list_out = FALSE)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{epi_names}{the name(s) of the output(s) to get (default = NULL)}

\item{list_out}{Should the output be a list containing the element(s)?
See **Value**. (default = FALSE)}
}
\value{
If `epi_names` is null, the full Epidemic Outputs list.
  If `epi_names` is of length one and `list_out == FALSE` (default),
  the `epi_names` elements of the Output list.
  If `list_out == TRUE`, a list containing the `out_name` element(s)
  of the parameters list.
}
\description{
Helper function to access the Epidemic Outputs list of the
             Master list object passed by modules during the course of a
             network simulation
}
\examples{

```
get_epi(dat)
get_epi(dat, "i")
get_epi(dat, "i", list_out = TRUE)
get_epi(dat, c("i", "s"), list_out = TRUE)
```

}
