% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{get_formula_terms}
\alias{get_formula_terms}
\title{Outputs Formula Terms into a Character Vector}
\usage{
get_formula_terms(formula)
}
\arguments{
\item{formula}{Right-hand sided formation formula.}
}
\description{
Given a formation formula for a network model, outputs it into
             a character vector of terms to be used in \code{netsim}
             simulations.
}
\seealso{
\code{\link{copy_toall_attr}}, \code{\link{get_attr_prop}},
         \code{\link{update_nwattr}}.
}
\keyword{internal}
\keyword{netUtils}
