% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.fn.modules.R
\name{get_modules}
\alias{get_modules}
\title{Get the List of Modules}
\usage{
get_modules(dat)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link[=netsim]{netsim()}}.}
}
\value{
A named list of modules to be run by the model
}
\description{
Get the List of Modules
}
\keyword{internal}
