% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{get_network_attributes}
\alias{get_network_attributes}
\title{Get Network Attributes from a Network Object}
\usage{
get_network_attributes(x)
}
\arguments{
\item{x}{An object of class \code{network} or \code{networkLite}.}
}
\value{
Returns the named list of network attributes.
}
\description{
Gets all network attributes except \code{"mnext"} from its
network argument.
}
\details{
This function is used in \code{EpiModel} workflows to copy relevant network
attributes from the network object to the \code{netsim_dat} object when
initializing \code{netsim} runs.
}
\examples{
nw <- network_initialize(100)
get_network_attributes(nw)

}
