% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_node_records}
\alias{get_node_records}
\title{Extract the node records from Network Simulations}
\usage{
get_node_records(sims)
}
\arguments{
\item{sims}{An \code{EpiModel} object of class \code{netsim}.}
}
\value{
A list of \code{data.frame}s, one for each "measure" recorded in the
simulation by the `record_node_value` function.
}
\description{
Extract the node records from Network Simulations
}
\details{
See the "Time Varying Attributes" vignette
}
