% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_nwparam.bip}
\alias{get_nwparam.bip}
\title{Extract Bipartite Network Model Parameters}
\usage{
get_nwparam.bip(x, network = 1)
}
\arguments{
\item{x}{Master data object used in \code{netsim.bip} simulations.}

\item{network}{Network number, for simulations with multiple networks
representing the population.}
}
\description{
Extracts a list of network model parameters saved in the
             initialization module.
}
\keyword{extract}
\keyword{internal}
