% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_nwparam}
\alias{get_nwparam}
\title{Extract Network Model Parameters}
\usage{
get_nwparam(x, network = 1)
}
\arguments{
\item{x}{master data object used in \code{netsim} simulations.}

\item{network}{network number, for simulations with multiple networks
       representing the population.}
}
\description{
Extracts a list of network model parameters saved in the
             initialization module.
}
\keyword{extract}
\keyword{internal}

