% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_nwstats.bip}
\alias{get_nwstats.bip}
\title{Extract Network Statistics from netsim.bip or netdx.bip Object}
\usage{
get_nwstats.bip(x, sim, network = 1)
}
\arguments{
\item{x}{An \code{EpiModel} object of class \code{\link{netsim.bip}} or
\code{\link{netdx.bip}}.}

\item{sim}{A vector of simulation numbers from the extracted object}

\item{network}{Network number, for \code{netsim.bip} objects with multiple
overlapping networks (advanced use, and not applicable to \code{netdx.bip}
objects).}
}
\description{
Extracts a data frame of network statistics from a network
             epidemic model simulated with \code{netsim.bip} or a network diagnostics
             object simulated with \code{netdx.bip}.
}
\examples{
# Bipartite Bernoulli random graph TERGM
nw <- network.initialize.bip(n = 100, bipartite = 50, directed = FALSE)
formation <- ~edges
target.stats <- 50
coef.diss <- dissolution_coefs.bip(dissolution = ~offset(edges), duration = 20)
est <- netest.bip(nw, formation, target.stats, coef.diss, verbose = FALSE)

dx <- netdx.bip(est, nsim = 3, nsteps = 10, verbose = FALSE,
            nwstats.formula = ~edges + isolates)
get_nwstats.bip(dx)
get_nwstats.bip(dx, sim = 1)

# SI epidemic model
param <- param.net.bip(inf.prob = 0.3, inf.prob.m2 = 0.15)
init <- init.net.bip(i.num = 10, i.num.m2 = 10)
control <- control.net.bip(type = "SI", nsteps = 10, nsims = 3,
                       nwstats.formula = ~edges + meandeg + degree(0:5),
                       verbose = FALSE)
mod <- netsim.bip(est, param, init, control)

# Extract the network statistics from all or sets of simulations
get_nwstats.bip(mod)
get_nwstats.bip(mod, sim = 2)
get_nwstats.bip(mod, sim = c(1, 3))

# On the fly summary stats
summary(get_nwstats.bip(mod))
colMeans(get_nwstats.bip(mod))

}
\keyword{extract}
