% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.dat.R
\name{get_param}
\alias{get_param}
\title{Get the parameters of a Master list object in network models}
\usage{
get_param(dat, param_names = NULL, list_out = FALSE)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{param_names}{the name(s) of the parameter(s) to get (default = NULL)}

\item{list_out}{Should the output be a list containing the element(s)?
See **Value**. (default = FALSE)}
}
\value{
If `param_names` is null, the full parameters list.
  If `param_names` is of length one and `list_out == FALSE` (default),
  the `param_names` element of the nodal param_ibute list.
  If `list_out == TRUE`, a list containing the `param_name` element(s)
  of the paramets list.
}
\description{
Helper function to access the parameters list of the
             Master list object passed by modules during the course of a
             network simulation
}
\examples{

```
get_param(dat)
get_param(dat, "inf.prob")
get_param(dat, "inf.prob", list_out = TRUE)
get_param(dat, c("inf.prob", "act.rate"), list_out = TRUE)
```

}
