% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.utils.R
\name{get_prev.icm.bip}
\alias{get_prev.icm.bip}
\title{Get Epidemic Output from icm Model}
\usage{
get_prev.icm.bip(dat, at)
}
\arguments{
\item{dat}{Master data list object.}

\item{at}{Current time step.}
}
\description{
This function provides all active model state sizes from
             the network at the specified time step, output to a list of
             vectors.
}
\keyword{internal}
