% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_prev.icm}
\alias{get_prev.icm}
\title{Get Epidemic Output from icm Model}
\usage{
get_prev.icm(dat, at)
}
\arguments{
\item{dat}{master data list object.}

\item{at}{current time step.}
}
\description{
This function provides all active model state sizes from
             the network at the specified time step, output to a list of
             vectors.
}
\keyword{internal}

