% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.getprev.R
\name{get_prev.net.grp}
\alias{get_prev.net.grp}
\title{Get Epidemic Output from netsim Model}
\usage{
get_prev.net.grp(dat, at)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{at}{Current time step.}
}
\description{
Provides all active model state sizes from the network at the
             specified time step, output to a list of vectors.
}
\details{
This network utility is used during the \code{\link{netsim}} simulation
process to efficiently query the current  state size for each mode and overall.
}
\keyword{internal}
\keyword{netUtils}
