% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_transmat.bip}
\alias{get_transmat.bip}
\title{Extract Transmissions Matrix from Bipartite Network Epidemic Model}
\usage{
get_transmat.bip(x, sim = 1)
}
\arguments{
\item{x}{An \code{EpiModel} object of class \code{\link{netsim.bip}}.}

\item{sim}{Simulation number of extracted network.}
}
\value{
A data frame with the following columns
\itemize{
 \item \strong{at:} the time step at which the transmission occurred.
 \item \strong{sus:} the ID number of the susceptible (newly infected) node.
 \item \strong{inf:} the ID number of the infecting node.
 \item \strong{infDur:} the duration of the infecting node's disease at the
       time of the transmission.
 \item \strong{transProb:} the probability of transmission per act.
 \item \strong{actRate:} the rate of acts per unit time.
 \item \strong{finalProb:} the final transmission probability for the
       transmission event.
}
}
\description{
Extracts the matrix of transmission data for each transmission
             event that occured within a bipartite network epidemic model.
}
\examples{
## Simulate SI epidemic on bipartite Bernoulli random graph
nw <- network.initialize.bip(n = 100, bipartite = 50, directed = FALSE)
formation <- ~edges
target.stats <- 50
coef.diss <- dissolution_coefs.bip(dissolution = ~offset(edges), duration = 20)
est <- netest.bip(nw, formation, target.stats, coef.diss, verbose = FALSE)
param <- param.net.bip(inf.prob = 0.3, inf.prob.m2 = 0.15)
init <- init.net.bip(i.num = 10, i.num.m2 = 10)
control <- control.net.bip(type = "SI", nsteps = 10, nsims = 3, verbose = FALSE)
mod <- netsim.bip(est, param, init, control)

## Extract the transmission matrix from simulation 2
get_transmat.bip(mod, sim = 2)

}
\keyword{extract}
