% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{groupids}
\alias{groupids}
\title{ID Numbers for Two-Group Network}
\usage{
groupids(nw, group)
}
\arguments{
\item{nw}{Object of class \code{network} or \code{networkDynamic}.}

\item{group}{Group number to return ID numbers for.}
}
\description{
Outputs ID numbers for a group number for a two-group network.
}
\examples{
nw <- network_initialize(n = 10)
nw <- set_vertex_attribute(nw, "group", rep(c(1,2), each = 5))
groupids(nw, group = 2)

}
\seealso{
\code{\link{idgroup}} provides the reverse functionality.
}
\keyword{internal}
\keyword{netUtils}
