% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{grpvals}
\alias{grpvals}
\title{Vertex Attributes for Two-Group Network}
\usage{
grpvals(nw, group, val)
}
\arguments{
\item{nw}{An object of class \code{network} or \code{networkDynamic}.}

\item{group}{group number.}

\item{val}{Nodal attribute to return.}
}
\description{
Outputs static vertex attributes for a two-group network for one
             specified group.
}
\examples{
nw <- network.initialize(n = 10)
nw <- set.vertex.attribute(nw, "male", rep(0:1, each = 5))
grpvals(nw, group = 1, "male")

}
\keyword{internal}
\keyword{netUtils}
