% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.R
\name{icm.bip}
\alias{icm.bip}
\title{Stochastic Bipartite Individual Contact Models}
\usage{
icm.bip(param, init, control)
}
\arguments{
\item{param}{Model parameters, as an object of class \code{\link{param.icm.bip}}.}

\item{init}{Initial conditions, as an object of class \code{\link{init.icm.bip}}.}

\item{control}{Control settings, as an object of class
\code{\link{control.icm.bip}}.}
}
\value{
A list of class \code{icm.bip} with the following elements:
\itemize{
 \item \strong{param:} the epidemic parameters passed into the model through
       \code{param.bip}, with additional parameters added as necessary.
 \item \strong{control:} the control settings passed into the model through
       \code{control.bip}, with additional controls added as necessary.
 \item \strong{epi:} a list of data frames, one for each epidemiological
       output from the model. Outputs for base models always include the
       size of each compartment, as well as flows in, out of, and between
       compartments.
}
}
\description{
Simulates stochastic bipartite individual contact epidemic models for
             infectious disease.
}
\details{
Individual contact models are intended to be the stochastic microsimulation
analogs to deterministic compartmental models. ICMs simulate disease spread
on individual agents in discrete time as a function of processes with stochastic
variation. The stochasticity is inherent in all transition processes:
infection, recovery, and demographics. A detailed description of these models
may be found in the \href{http://statnet.github.io/tut/BasicICMs.html}{Basic
ICMs} tutorial.

The \code{icm.bip} function performs  modeling of both the base model types
and original models. Base model types include two-group models with disease
types for Susceptible-Infected (SI), Susceptible-Infected-Recovered (SIR),
and Susceptible-Infected-Susceptible (SIS).
Original models may be built by writing new process modules that either take
the place of existing modules (for example, disease recovery), or supplement
the set of existing processes with a new one contained in an original module.
}
\examples{
\dontrun{
## Example: SI Model with Vital Dynamics (Two-Group)
param <- param.icm.bip(inf.prob = 0.4,  inf.prob.g2 = 0.1,
                   act.rate = 0.25, balance = "g1",
                   a.rate = 1/100, a.rate.g2 = NA,
                   ds.rate = 1/100, ds.rate.g2 = 1/100,
                   di.rate = 1/50, di.rate.g2 = 1/50)
init <- init.icm.bip(s.num = 500, i.num = 1,
                 s.num.g2 = 500, i.num.g2 = 0)
control <- control.icm.bip(type = "SI", nsteps = 500, nsims = 10)
mod4 <- icm.bip(param, init, control)
mod4
plot(mod4)
}

}
\seealso{
Extract the model results with \code{\link{as.data.frame.icm.bip}}.
Summarize the time-specific model results with \code{\link{summary.icm.bip}}.
Plot the model results with \code{\link{plot.icm.bip}}. Plot a compartment flow
diagram with \code{\link{comp_plot.bip}}.
}
\keyword{model}
