% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{idmode.bip}
\alias{idmode.bip}
\title{Mode Numbers for Bipartite Network}
\usage{
idmode.bip(nw, ids)
}
\arguments{
\item{nw}{Object of class \code{network.bip} or \code{networkDynamic.bip}.}

\item{ids}{Vector of ID numbers for which the mode number
should be returned.}
}
\description{
Outputs mode numbers give ID numbers for a bipartite network.
}
\examples{
nw <- network.initialize.bip(10, bipartite = 5)
idmode.bip(nw)
idmode.bip(nw, ids = c(3, 6))

}
\seealso{
\code{\link{modeids.bip}} provides the reverse functionality.
}
\keyword{internal}
\keyword{netUtils}
