% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.helpers.R
\name{increment_timestep}
\alias{increment_timestep}
\title{Increment the Current Timestep}
\usage{
increment_timestep(dat)
}
\arguments{
\item{dat}{Main data object passed through \code{netsim} simulations.}
}
\value{
A \code{dat} main list object.
}
\description{
This function adds 1 to the timestep counter stored in the
             \code{dat} main list object.
}
\section{Mutability}{

This DOES NOT modify the \code{dat} object in place. The result must be
assigned back to \code{dat} in order to be registered:
\code{dat <- increment_timestep(dat)}.
}

