% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/icm.mod.status.R
\name{infection.icm}
\alias{infection.icm}
\title{Primary Infection Module for icm}
\usage{
infection.icm(dat, at)
}
\arguments{
\item{dat}{Master data list object.}

\item{at}{Current time step.}
}
\description{
This function simulates the main infection process given the
             current state of the actors in the system.
}
\keyword{internal}

