% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init.icm.bip}
\alias{init.icm.bip}
\title{Initial Conditions for Stochastic Bipartite Individual Contact Models}
\usage{
init.icm.bip(s.num, i.num, r.num, s.num.g2, i.num.g2, r.num.g2, ...)
}
\arguments{
\item{s.num}{Number of initial group 1 susceptible.}

\item{i.num}{Number of initial group 1 infected.}

\item{r.num}{Number of initial group 1 recovered. This parameter is only used for
the \code{SIR} model type.}

\item{s.num.g2}{Number of initial susceptible in group 2.}

\item{i.num.g2}{Number of initial infected in group 2.}

\item{r.num.g2}{Number of initial recovered in group 2. This parameter is
only used for two-group \code{SIR} models.}

\item{...}{Additional initial conditions passed to model.}
}
\description{
Sets the initial conditions for stochastic individual contact
             models simulated with \code{icm}.
}
\details{
The initial conditions for a model solved with \code{\link{icm.bip}} should be
input into the \code{init.icm.bip} function. This function handles initial
conditions for both base models and original models using new modules. For
an overview of initial conditions for base ICM class models, consult the
\href{http://statnet.github.io/tut/BasicICMs.html}{Basic ICMs} tutorial.
}
\seealso{
Use \code{\link{param.icm.bip}} to specify model parameters and
         \code{\link{control.icm.bip}} to specify the control settings. Run the
         parameterized model with \code{\link{icm.bip}}.
}
\keyword{parameterization}
