% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.mod.init.R
\name{init_pids}
\alias{init_pids}
\title{Persistent ID Initialization}
\usage{
init_pids(nw, prefixes = c("F", "M"))
}
\arguments{
\item{nw}{An object of class \code{networkDynamic}.}

\item{prefixes}{Character string prefix for mode-specific ID.}
}
\description{
This function initializes the persistent IDs for
             a \code{networkDynamic} object.
}
\details{
This function is used for \code{\link{netsim}} simulations
over bipartite networks for populations with vital dynamics. Persistent IDs are
required in this situation because when new nodes are added to the
first mode in a bipartite network, the IDs for the second mode shift
upward. Persistent IDs allow for an analysis of disease transmission
chains for these simulations. These IDs are also invoked in the
\code{\link{births.net}} module when the persistent IDs of incoming nodes
must be set.
}
\examples{
# Initialize network with 25 female and 75 male
nw <- network.initialize(100, bipartite = 25)

# Set persistent IDs using the default F/M prefix
nw <- init_pids(nw)
nw \%v\% "vertex.names"

# Use another prefix combination
nw <- init_pids(nw, c("A", "B"))
nw \%v\% "vertex.names"
}
\seealso{
\code{\link{initialize.pids}}
}
\keyword{internal}
\keyword{netMod}

