% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.mod.init.R
\name{init_status.icm.bip}
\alias{init_status.icm.bip}
\title{Disease Status Initialization Module for bipartite icm}
\usage{
init_status.icm.bip(dat)
}
\arguments{
\item{dat}{Master data list object.}
}
\description{
This function sets the initial disease status on the
             network given the specified initial conditions.
}
\seealso{
This is an initialization module for \code{\link{icm.bip}}.
}
\keyword{internal}
