% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.mod.init.R
\name{init_status.icm}
\alias{init_status.icm}
\title{Disease Status Initialization Module for icm}
\usage{
init_status.icm(dat)
}
\arguments{
\item{dat}{Main \code{icm_dat} class data object passed through \code{icm}
simulations.}
}
\value{
The updated \code{icm_dat} class main data object.
}
\description{
This function sets the initial disease status on the
network given the specified initial conditions.
}
\seealso{
This is an initialization module for \code{\link[=icm]{icm()}}.
}
\keyword{internal}
