% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErgmTerms.R
\name{InitErgmTerm.absdiffby}
\alias{InitErgmTerm.absdiffby}
\alias{absdiffby}
\title{Definition for absdiffby ERGM Term}
\usage{
InitErgmTerm.absdiffby(nw, arglist, ...)
}
\arguments{
\item{nw}{An object of class \code{network}.}

\item{arglist}{A list of arguments as specified in the \code{ergm.userterms}
package framework.}

\item{...}{Additional data passed into the function as specified in the
\code{ergm.userterms} package framework.}
}
\description{
This function defines and initializes the absdiffby ERGM term
             that allows for representing homophily with respect to a
             non-binary attribute (e.g., age) differentially by a binary
             attribute (e.g., sex).
}
\details{
This ERGM user term was written to allow for age-based homophily in
partnership formation that is asymmetric by sex. The \code{absdiff} component
targets age-based homophily while the \code{by} component allows that to be
structured by a binary attribute such as "male", in order to enforce an
offset in the average difference. This allows, for example, a average age
difference in partnerships, but with males (on average) older than females.
}
