% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.mod.init.R
\name{initialize.icm.bip}
\alias{initialize.icm.bip}
\title{Initialization: bipartite icm Module}
\usage{
initialize.icm.bip(param, init, control)
}
\arguments{
\item{param}{An \code{EpiModel} object of class \code{\link{param.icm.bip}}.}

\item{init}{An \code{EpiModel} object of class \code{\link{init.icm.bip}}.}

\item{control}{An \code{EpiModel} object of class \code{\link{control.icm.bip}}.}
}
\description{
This function initializes the master \code{dat} object on which
             data are stored, and simulates disease status and other attributes.
}
\keyword{internal}
