% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.mod.init.R
\name{initialize.icm}
\alias{initialize.icm}
\title{Initialization: icm Module}
\usage{
initialize.icm(param, init, control)
}
\arguments{
\item{param}{An `EpiModel` object of class `\link{param.icm}`.}

\item{init}{An `EpiModel` object of class `\link{init.icm}`.}

\item{control}{An `EpiModel` object of class `\link{control.icm}`.}
}
\description{
This function initializes the master `dat` object on which
             data are stored, and simulates disease status and other attributes.
}
\keyword{internal}
