% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.accessor.R
\name{is_active_unique_ids}
\alias{is_active_unique_ids}
\title{Are These Nodes Active (Unique IDs)}
\usage{
is_active_unique_ids(dat, unique_ids)
}
\arguments{
\item{dat}{Main data object passed through \code{netsim} simulations.}

\item{unique_ids}{A vector of node unique identifiers.}
}
\value{
A logical vector with TRUE if the node is still active and FALSE
otherwise.
}
\description{
Are These Nodes Active (Unique IDs)
}
