% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdx.R
\name{make_dissolution_stats}
\alias{make_dissolution_stats}
\title{Calculate the Dissolution Statistics of a Network}
\usage{
make_dissolution_stats(diag.sim, coef.diss, nsteps, verbose = TRUE)
}
\arguments{
\item{diag.sim}{A list of network objects (one per simulation).}

\item{coef.diss}{The \code{coef.diss} element of \code{nwparam}.}

\item{nsteps}{The number of simulated steps.}

\item{verbose}{A verbosity toggle (default = TRUE).}
}
\value{
a \code{list} of dissolution statistics
}
\description{
Calculate the Dissolution Statistics of a Network
}
\keyword{internal}
