% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.fn.modules.R
\name{make_module_list}
\alias{make_module_list}
\title{Populate the Module List After the Initialization Module is run}
\usage{
make_module_list(dat)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link[=netsim]{netsim()}}.}
}
\value{
The updated \code{netsim_dat} main list object.
}
\description{
Populate the Module List After the Initialization Module is run
}
\keyword{internal}
