% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.simnet.R
\name{make_sim_network}
\alias{make_sim_network}
\title{Construct Network from dat Object}
\usage{
make_sim_network(dat, network = 1L)
}
\arguments{
\item{dat}{Main list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}

\item{network}{index of the network to construct}
}
\value{
The network.
}
\description{
This function returns the network object representing the
             current state of the simulation.
}
\keyword{internal}
\keyword{netUtils}
