% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdx.R
\name{make_stats_table}
\alias{make_stats_table}
\title{Create a Summary Table of Simulation Statistics}
\usage{
make_stats_table(stats, targets)
}
\arguments{
\item{stats}{A list of simulated statistics matrices, of length equal to the
number of simulations performed.  Each matrix should have one row for each
simulated network if \code{dynamic == FALSE}, one row for each time step 
if \code{dynamic == TRUE}, and one column for each statistic.}

\item{targets}{A vector of target values for the statistics in \code{stats}.
May be named (in which case targets will be matched to statistics based on
column names in matrices in \code{stats}) or unnamed (in which case 
targets will be matched to statistics based on position).}
}
\value{
A \code{data.frame} summarizing the simulated statistics.
}
\description{
Create a Summary Table of Simulation Statistics
}
\keyword{internal}
