% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{mcat}
\alias{mcat}
\title{Write Out Test Progress to Console}
\usage{
mcat(test)
}
\arguments{
\item{test}{Character string with the name of a test.}
}
\description{
Writes the name of a test and \code{...} to console to show test
             progress.
}
\keyword{internal}
