% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{modeids.bip}
\alias{modeids.bip}
\title{ID Numbers for Bipartite Network}
\usage{
modeids.bip(nw, mode)
}
\arguments{
\item{nw}{Object of class \code{network.bip} or \code{networkDynamic.bip}.}

\item{mode}{Mode number to return ID numbers for.}
}
\description{
Outputs ID numbers for a mode number for a bipartite network.
}
\examples{
nw <- network.initialize.bip(10, bipartite = 5)
modeids.bip(nw, mode = 2)

}
\seealso{
\code{\link{idmode.bip}} provides the reverse functionality.
}
\keyword{internal}
\keyword{netUtils}
