% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.fn.modules.R
\name{modules_updater}
\alias{modules_updater}
\title{Function to Modify the List of Modules During the Simulation}
\usage{
modules_updater(dat)
}
\arguments{
\item{dat}{Main list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}
}
\value{
The updated \code{dat} main list object.
}
\description{
Function to Modify the List of Modules During the Simulation
}
\details{
This function removes modules from the module list by reading the
".modules.to.remove" parameter. If the parameter is empty, the function
immediatly returns. Otherwise, the modules specified are removed. After the
removal, the ".modules.to.remove" parameter is set back to NULL.
Therefore, this functionnality is to be used with the "scenarios" API or the
".param.updater" parameter as module removal should happen at specified
timestep.
}
\keyword{internal}
