% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_epi.bip}
\alias{mutate_epi.bip}
\title{Add New Epidemiology Variables}
\usage{
mutate_epi.bip(x, ...)
}
\arguments{
\item{x}{An \code{EpiModel} object of class \code{dcm.bip}, \code{icm.bip}, or
\code{netsim.bip}.}

\item{...}{Name-value pairs of expressions (see examples below).}
}
\description{
Inspired by \code{dplyr::mutate}, \code{mutate_epi.bip} adds new
             variables to the epidemiological and related variables within
             simulated model objects of any class in \code{EpiModel}.
}
\examples{
# DCM example
param <- param.dcm.bip(inf.prob = 0.2, act.rate = 0.25)
init <- init.dcm.bip(s.num = 500, i.num = 1)
control <- control.dcm.bip(type = "SI", nsteps = 500)
mod1 <- dcm.bip(param, init, control)
mod1 <- mutate_epi(mod1, prev = i.num/num)
plot(mod1, y = "prev")

# Network model example
nw <- network.initialize.bip(n = 100, bipartite = 50, directed = FALSE)
formation <- ~edges
target.stats <- 50
coef.diss <- dissolution_coefs.bip(dissolution = ~offset(edges), duration = 20)
est1 <- netest.bip(nw, formation, target.stats, coef.diss, verbose = FALSE)

param <- param.net.bip(inf.prob = 0.3, inf.prob.m2 = 0.15)
init <- init.net.bip(i.num = 1, i.num.m2 = 0)
control <- control.net.bip(type = "SI", nsteps = 10, nsims = 3,
                       verbose = FALSE)
mod1 <- netsim.bip(est1, param, init, control)
mod1

# Add the prevalences to the dataset
mod1 <- mutate_epi(mod1, i.prev = i.num / num,
                         i.prev.m2 = i.num.m2 / num.m2)
plot(mod1, y = c("i.prev", "i.prev.m2"), qnts = 0.5, legend = TRUE)

# Add incidence rate per 100 person years (assume time step = 1 week)
mod1 <- mutate_epi(mod1, ir100 = 5200*(si.flow + si.flow.m2) /
                                      (s.num + s.num.m2))
as.data.frame(mod1)
as.data.frame(mod1, out = "mean")

}
