% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{netsim_cond_msg}
\alias{netsim_cond_msg}
\title{Message to find in which module a `condition` occured}
\usage{
netsim_cond_msg(cond, module, at, msg)
}
\arguments{
\item{cond}{The type of `condition` handled (message, warning, error)}

\item{module}{The name of the module where the `condition` occured}

\item{at}{The time step the `condition` occured}

\item{msg}{The `condition`'s message}
}
\value{
A formatted string describing where and whe the `condition` occured
        as well as the `condition`'s message.
}
\description{
Returns vector of colors consistent with a high-brightness set
             of colors from an \code{RColorBrewer} palette.
}
\keyword{internal}
