% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{netsim_error_logger}
\alias{netsim_error_logger}
\title{Handle the Logging of Traceback and Dumping of Frames on Error}
\usage{
netsim_error_logger(dat, s)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link[=netsim]{netsim()}}.}

\item{s}{The number of the simulation that failed}
}
\value{
Nothing, after logging and dumping frames, the function gives the
control back to the general error handler
}
\description{
If \code{control$.traceback.on.error == TRUE}, this function prints the traceback
of the current simulation to STDIN. This is useful when \code{ncores > 1} or in
HPC settings.
If \code{control$.dump.frames.on.error == TRUE}, this function saves a debugging
dump for "postmortem debugging". The dumps are named
"dump_\%Y\%m\%d_\%H\%M\%S_s.rda" and stored at the root of the working directory.
}
\keyword{internal}
