% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsim.R
\name{netsim_loop}
\alias{netsim_loop}
\title{Internal function running the network simulation loop}
\usage{
netsim_loop(x, param, init, control, s)
}
\arguments{
\item{x}{An \code{EpiModel} object of class \code{\link{netest}}.}

\item{param}{An \code{EpiModel} object of class \code{\link{param.net}}.}

\item{init}{An \code{EpiModel} object of class \code{\link{init.net}}.}

\item{control}{An \code{EpiModel} object of class \code{\link{control.net}}.}

\item{s}{Simulation number, used for restarting dependent simulations.}
}
\description{
This function run the initialization and simulation loop for one
             simulation
}
\keyword{internal}
