% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{networkLitemethods}
\alias{networkLitemethods}
\alias{get.vertex.attribute.networkLite}
\alias{set.vertex.attribute.networkLite}
\alias{list.vertex.attributes.networkLite}
\alias{get.network.attribute.networkLite}
\alias{set.network.attribute.networkLite}
\alias{list.network.attributes.networkLite}
\alias{network.edgecount.networkLite}
\alias{as.edgelist.networkLite}
\alias{mixingmatrix.networkLite}
\alias{[<-.networkLite}
\alias{print.networkLite}
\alias{network.naedgecount.networkLite}
\alias{add.edges.networkLite}
\alias{as.networkLite}
\alias{as.networkLite.network}
\alias{as.networkLite.networkLite}
\title{networkLite Methods}
\usage{
\method{get.vertex.attribute}{networkLite}(x, attrname, ...)

\method{set.vertex.attribute}{networkLite}(x, attrname, value, v = seq_len(network.size(x)), ...)

\method{list.vertex.attributes}{networkLite}(x, ...)

\method{get.network.attribute}{networkLite}(x, attrname, ...)

\method{set.network.attribute}{networkLite}(x, attrname, value, ...)

\method{list.network.attributes}{networkLite}(x, ...)

\method{network.edgecount}{networkLite}(x, ...)

\method{as.edgelist}{networkLite}(x, output = c("matrix", "tibble"), ...)

\method{mixingmatrix}{networkLite}(object, attr, ...)

\method{[}{networkLite}(x, i, j) <- value

\method{print}{networkLite}(x, ...)

\method{network.naedgecount}{networkLite}(x, ...)

\method{add.edges}{networkLite}(
  x,
  tail,
  head,
  names.eval = NULL,
  vals.eval = NULL,
  ...,
  check.unique = FALSE
)

as.networkLite(x, ...)

\method{as.networkLite}{network}(x, ...)

\method{as.networkLite}{networkLite}(x, ...)
}
\arguments{
\item{x}{A \code{networkLite} object.}

\item{attrname}{The name of an attribute in \code{x}.}

\item{...}{Any additional arguments.}

\item{value}{The attribute value to set for
\code{set.vertex.attribute.networkLite}; the value to set edges
to (must be FALSE) for the \code{networkLite} replacement
method.}

\item{v}{Indices at which to set vertex attribute values.}

\item{output}{Type of edgelist to output.}

\item{object}{A \code{networkLite} object.}

\item{attr}{Specification of a vertex attribute in \code{object} as
described in \code{\link[ergm]{nodal_attributes}}.}

\item{i, j}{Nodal indices (must be missing for networkLite method).}

\item{tail}{Vector of tails of edges to add to the networkLite.}

\item{head}{Vector of heads of edges to add to the networkLite.}

\item{names.eval}{Currently unsupported by add.edges.networkLite.}

\item{vals.eval}{Currently unsupported by add.edges.networkLite.}

\item{check.unique}{Should a check to ensure uniqueness of edges
in the final edgelist be performed?}
}
\description{
S3 methods for networkLite class, for generics defined in
             network package.
}
\details{
Allows use of networkLite objects in \code{ergm_model}.
}
