% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.accessor.R
\name{padded_vector}
\alias{padded_vector}
\title{Grow a Vector to a Given Size, Padding it With Empty Elements}
\usage{
padded_vector(orig, size)
}
\arguments{
\item{orig}{a vector to grow}

\item{size}{the final size of the vector}
}
\value{
a vector of size \code{size} padded with \code{NULL}s or \code{NA}s at the end
}
\description{
Grow a vector to a given size, padding it with  \code{NULL} if \code{orig} is a \code{list}
and with \code{NA} otherwise
}
