% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.inputs.R
\name{param.net_from_table}
\alias{param.net_from_table}
\title{Parameters List for Stochastic Network Models from a Formatted Data Frame}
\usage{
param.net_from_table(long.param.df)
}
\arguments{
\item{long.param.df}{A \code{data.frame} of parameters. See details for the
expected format.}
}
\value{
A list object of class \code{param.net}, which can be passed to
  EpiModel function \code{\link{netsim}}.
}
\description{
Sets the epidemic parameters for stochastic network models with
\code{\link{netsim}} using a specially formatted data frame of parameters.
}
\details{
The \code{long.param.df} first 3 columns must be:
1. 'param': the name of the parameter. If it is a multi dimension parameter,
   prepend the name, with "_1", "_2", etc for the position in the vector.
2. 'value': the value for the parameter (or the Nth position if
multidimensional)
3. 'type': a character string containing either "numeric", "logical" or
   "character". The type the value will be cast to.

Apart from these 3 columns, the \code{data.frame} can contain any number
of other columns. A typical use case would be to have a "details" and
"source" columns to document where these parameters come from.
}
