% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendo.R
\name{plot.transmat}
\alias{plot.transmat}
\title{Plot transmat infection tree in one of several styles}
\usage{
\method{plot}{transmat}(x, style = c("phylo", "network", "transmissionTimeline"), ...)
}
\arguments{
\item{x}{A `\link{transmat}` object to be plotted}

\item{style}{Character name of plot style. One of "phylo", "network",
or "transmissionTimeline"}

\item{...}{Additional plot arguments to be passed to lower-level plot
functions (plot.network, plot.phylo, etc)}
}
\description{
Plots the infection tree described in a `\link{transmat}`
             object in one of several styles: phylogenetic tree, an un-rooted
             network, a hierarchical tree, or a transmissionTimeline.
}
\details{
The phylo plot requires the `ape` package. The
`ndtv::transmissionTimeline` requires that the `ndtv` package
is installed. All of the options are essentially
wrappers to other plot calls with some appropriate preset arguments.
}
\seealso{
`\link{plot.network}`,`\link[ape]{plot.phylo}`
}
