% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.prevalence.R
\name{prevalence.net}
\alias{prevalence.net}
\title{Get Epidemic Output from netsim Model}
\usage{
prevalence.net(dat, at)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}

\item{at}{Current time step.}
}
\description{
Provides all active model state sizes from the network at the
             specified time step, output to a list of vectors.
}
\details{
This network utility is used during the \code{\link{netsim}} simulation
process to efficiently query the current size of each state or compartment
in the model at any given timestep. For a two-group network, the current
state size for each group and overall is provided.
}
\keyword{internal}
\keyword{netUtils}
