% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print_nwstats_table}
\alias{print_nwstats_table}
\title{Print Helper For Network Stats Tables}
\usage{
print_nwstats_table(nwtable, digits)
}
\arguments{
\item{nwtable}{a formation or dissolution statistics \code{data.frame}}

\item{digits}{argument to be passed to \code{round}}
}
\description{
Print Helper For Network Stats Tables
}
