% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.record.R
\name{record_node_value}
\alias{record_node_value}
\title{Record time varying node values}
\usage{
record_node_value(dat, at, measure, posit_ids, values)
}
\arguments{
\item{dat}{a Master list object of network models}

\item{at}{the time step where the recording happens}

\item{measure}{the name of the value to record}

\item{posit_ids}{a numeric vector of posit_ids to which the measure applies}

\item{values}{the values to be recorded}
}
\value{
The Master list object


See the "Time Varying Attributes" vignette
}
\description{
This function records values specific to a time step and a group of nodes.
The nodes are identified by their \code{unique_ids} which allows the
recording of data for nodes that are no longer in the network by the end of
the run. The records are stored in \code{dat$node.records} and can be
accessed from the \code{netsim} object with \code{get_node_records}
}
\examples{
\dontrun{

dat <- record_node_value(dat, at, "attr_1", get_posit_ids(dat), 5)
some_nodes <- get_posit_ids(dat)
some_nodes <- some_nodes[runif(length(some_nodes)) < 0.2]
dat <- record_node_value(
  dat, at,
  "attr_2",
  some_nodes,
  rnorm(length(some_nodes))
)

}

}
