% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.mod.status.R
\name{recovery.icm.bip}
\alias{recovery.icm.bip}
\title{Recovery: icm Module}
\usage{
recovery.icm.bip(dat, at)
}
\arguments{
\item{dat}{Main data object passed through \code{icm} simulations.}

\item{at}{Current time step.}
}
\value{
The updated \code{netsim_dat} main list object.
}
\description{
This function simulates recovery from the infected state
either to a distinct recovered state (SIR model type) or back
to a susceptible state (SIS model type), for use in
\code{\link{icm}}.
}
\keyword{internal}
