% Generated by roxygen2 (4.0.2): do not edit by hand
\name{recovery.net}
\alias{recovery.net}
\title{Recovery: netsim Module}
\usage{
recovery.net(all, at)
}
\arguments{
\item{all}{a list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{at}{current time step.}
}
\description{
This function simulates recovery from the infected state
             either to an distinct recovered state (SIR model type) or back
             to a susceptible state (SIS model type), for use in
             \code{\link{netsim}}.
}
\keyword{internal}

