% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModel-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{control.ergm.ego}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ergm.ego}{\code{\link[ergm.ego]{control.ergm.ego}}}
}}

