% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.fn.modules.R
\name{remove_modules}
\alias{remove_modules}
\title{Remove a Set of Modules From the Module List}
\usage{
remove_modules(dat, names.to.remove)
}
\arguments{
\item{dat}{Main list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}

\item{names.to.remove}{a character vector containing the name of the modules
to remove.}
}
\value{
The updated \code{dat} main list object.
}
\description{
Remove a Set of Modules From the Module List
}
\keyword{internal}
