% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.simnet.R
\name{resim_nets.bip}
\alias{resim_nets.bip}
\title{Resimulate Dynamic Network at Time 2+}
\usage{
resim_nets.bip(dat, at)
}
\arguments{
\item{at}{Current time step.}

\item{x}{A master object passed through \code{\link{netsim.bip}}.}
}
\description{
This function resimulates the dynamic network in stochastic
             network models simulated in \code{\link{netsim.bip}} with dependence
             between the epidemic and demographic processes and the network
             structure.
}
\keyword{internal}
\keyword{netUtils}
