% Generated by roxygen2 (4.0.1): do not edit by hand
\name{saveout.dcm}
\alias{saveout.dcm}
\title{Save dcm Data to Output List Format}
\usage{
saveout.dcm(df, s, param, control, out)
}
\arguments{
\item{all}{master object in \code{df} simulations.}

\item{s}{current run number.}

\item{param}{param list set in \code{\link{param.dcm}}}

\item{control}{control list set in \code{\link{control.dcm}}.}

\item{out}{out list passed back in for updating at runs 2+.}
}
\description{
This function transfers the data from the master \code{df}
             object to the output \code{out} object at the end of each
             simulation in \code{\link{dcm}}.
}
\keyword{internal}

