% Generated by roxygen2 (4.0.2): do not edit by hand
\name{saveout.icm}
\alias{saveout.icm}
\title{Save icm Data to Output List Format}
\usage{
saveout.icm(dat, s, out)
}
\arguments{
\item{s}{current run number.}

\item{out}{out list passed back in for updating at runs 2+.}

\item{all}{master object in \code{icm} simulations.}
}
\description{
This function transfers the data from the master \code{all}
             object to the output \code{out} object at the end of each
             simulation in \code{\link{icm}}.
}
\keyword{internal}

