% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.dat.R
\name{set_attr}
\alias{set_attr}
\title{Edit an element of the nodale attributes of a Master list object}
\usage{
set_attr(dat, attr_name, value)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{attr_name}{the name of the attribute to modify}

\item{value}{the new value for `attr_name`}
}
\value{
The modified Master list object
}
\description{
Helper function to modify the nodale attribute list of the
             Master list object passed by modules during the course of a
             network simulation
}
\examples{

```
dat <- set_attr(dat, "age", new_ages)
```

}
\keyword{Internal}
