% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.accessor.R
\name{set_attr_full}
\alias{set_attr_full}
\title{Version of \code{set_attr} without indexes
keywords internal}
\usage{
set_attr_full(dat, item, value, override.length.check)
}
\description{
Version of \code{set_attr} without indexes
keywords internal
}
