% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.helpers.R
\name{set_current_timestep}
\alias{set_current_timestep}
\title{Sets the Current Timestep}
\usage{
set_current_timestep(dat, timestep)
}
\arguments{
\item{dat}{a Master list object of network models}

\item{timestep}{the new value for the timestep}
}
\value{
A \code{dat} Master list object
}
\description{
Changes the current timestep in the \code{dat} object. Use with
             caution. This function exists to workaround unforseen corner
             cases. In most situation, \code{increment_timestep} should be
             prefered
}
\section{Mutability}{

This DOES NOT modify the dat object in place. The result must be assigned
back to \code{dat} in order to be registered
\code{dat <- increment_timestep(dat)}
}

