% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.dat.R
\name{set_dat_elt}
\alias{set_dat_elt}
\title{internal helper function to edit elements of a Master list object}
\usage{
set_dat_elt(dat, elt, attrname, value)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{elt}{the name of the sublist to access}

\item{attrname}{the name of the output(s) to get (default = NULL)}
}
\value{
The modified Master list object
}
\description{
internal helper function to edit elements of a Master list object
}
\examples{

```
dat <- set_dat_elt(dat, "attr", "age", new_ages)
```

}
\keyword{Internal}
