% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.fn.modules.R
\name{set_modules}
\alias{set_modules}
\title{Set the List of Modules}
\usage{
set_modules(dat, modules)
}
\arguments{
\item{dat}{Main list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}

\item{modules}{A named list of modules to be run by the model}
}
\value{
The updated \code{dat} main list object.
}
\description{
Set the List of Modules
}
\keyword{internal}
