% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.dat.R
\name{set_param}
\alias{set_param}
\title{Edit an element of the parameters of a Master list object}
\usage{
set_param(dat, param_name, value)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{param_name}{the name of the parameter to modify}

\item{value}{the new value for `param_name`}
}
\value{
The modified Master list object
}
\description{
Helper function to modify the parameters list of the
             Master list object passed by modules during the course of a
             network simulation
}
\examples{

```
dat <- set_param(dat, "act.rate", new_rate)
```

}
\keyword{Internal}
