% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.simnet.R
\name{set_sim_network}
\alias{set_sim_network}
\title{Set Network on dat Object}
\usage{
set_sim_network(dat, nw)
}
\arguments{
\item{dat}{Main list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}

\item{nw}{the network}
}
\value{
The updated \code{dat} main list object.
}
\description{
This function updates the dat object given the network
             representing the current state of the simulation.
}
\keyword{internal}
\keyword{netUtils}
