% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.infection.R
\name{set_transmat}
\alias{set_transmat}
\title{Save Transmission Matrix}
\usage{
set_transmat(dat, del, at)
}
\arguments{
\item{dat}{Master list object containing a `networkDynamic` object or
edgelist (if tergmLite is used) and other initialization information
passed from `\link{netsim}`.}

\item{del}{Discordant edgelist created within `infection.net` and
`infection.2g.net`.}

\item{at}{Current time step.}
}
\description{
This function appends the transmission matrix created during
             `infection.net` and `infection.2g.net`.
}
\details{
This internal function works within the parent `\link{infection.net}` functions
to save the transmission matrix created at time step `at` to the master list object
`dat`.
}
