% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.simnet.R
\name{sim_nets.bip}
\alias{sim_nets.bip}
\title{Simulate Dynamic Network at Time 1}
\usage{
sim_nets.bip(x, nw, nsteps, control)
}
\arguments{
\item{x}{An \code{EpiModel} object of class \code{\link{netest.bip}}.}

\item{nw}{A \code{networkDynamic.bip} object.}

\item{nsteps}{Number of time steps to simulate the network over.}

\item{control}{An \code{EpiModel} object of class \code{\link{control.net.bip}}.}

\item{nsims}{Number of independent network simulations.}
}
\description{
This function simulates a dynamic network over one or multiple
             time steps, to be used in \code{\link{netsim.bip}} models.
}
\keyword{internal}
\keyword{netUtils}
