% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.simnet.R
\name{sim_nets}
\alias{sim_nets}
\title{Simulate Dynamic Network at Time 1}
\usage{
sim_nets(x, nw, nsteps, control)
}
\arguments{
\item{x}{An `EpiModel` object of class `\link{netest}`.}

\item{nw}{A `networkDynamic` object.}

\item{nsteps}{Number of time steps to simulate the network over.}

\item{control}{An `EpiModel` object of class `\link{control.net}`.}

\item{nsims}{Number of independent network simulations.}
}
\description{
This function simulates a dynamic network over one or multiple
             time steps, to be used in `\link{netsim}` models.
}
\keyword{internal}
\keyword{netUtils}
