% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.simnet.R
\name{simulate_dat}
\alias{simulate_dat}
\title{Simulate a Network for a Specified Number of Time Steps}
\usage{
simulate_dat(dat, at, nsteps = 1L)
}
\arguments{
\item{dat}{Main list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}

\item{at}{Current time step.}

\item{nsteps}{number of time steps to simulate}
}
\value{
The updated \code{dat} main list object.
}
\description{
This function simulates a dynamic network over one or multiple
             time steps for TERGMs or one or multiple cross-sectional network
             panels for ERGMs, for use in \code{\link{netsim}} modeling.
             Network statistics are also extracted and saved if 
             \code{save.nwstats == TRUE}.
}
\keyword{internal}
\keyword{netUtils}
