% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{snctrl}
\alias{snctrl}
\title{Statnet Control}
\description{
A utility to facilitate argument completion of control lists,
reexported from `statnet.common`.
}
\seealso{
[statnet.common::snctrl()]
}
