% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.icm.bip}
\alias{summary.icm.bip}
\title{Summary Model Statistics}
\usage{
\method{summary}{icm}(object, at, digits = 3, ...)
}
\arguments{
\item{object}{An \code{EpiModel} object of class \code{icm.bip}.}

\item{at}{Time step for model statistics.}

\item{digits}{Number of significant digits to print.}

\item{...}{Additional summary function arguments.}
}
\description{
Extracts and prints model statistics simulated with \code{icm.bip}.
}
\details{
Summary statistics for the main epidemiological outcomes (state and transition
size and prevalence) from an \code{icm.bip} model. Time-specific summary measures
are provided, so it is necessary to input a time of interest.
}
\examples{
## Stochastic ICM SI model with 3 simulations (Need Example)
param <- param.icm.bip(inf.prob = 0.2, act.rate = 1)
init <- init.icm.bip(s.num = 500, i.num = 1)
control <- control.icm.bip(type = "SI", nsteps = 50,
                       nsims = 5, verbose = FALSE)
mod <- icm.bip(param, init, control)
summary(mod, at = 25)
summary(mod, at = 50)

}
\seealso{
\code{\link{icm.bip}}
}
\keyword{extract}
