% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.netest}
\alias{summary.netest}
\title{Summary for Network Model Fit}
\usage{
\method{summary}{netest}(object, ...)
}
\arguments{
\item{object}{an \code{EpiModel} object of class \code{netest}.}

\item{...}{additional summary function arguments.}
}
\description{
Prints the summary model fit statistics for an ERGM or STERGM fit.
}
\details{
This function is simply a wrapper function for \code{summary.ergm} and
  \code{summary.stergm}.
}
\keyword{internal}

