% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.summary.R
\name{summary_nets}
\alias{summary_nets}
\title{Extract Summary Statistics of Networks Used in netsim}
\usage{
summary_nets(dat, at)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{netsim}}.}

\item{at}{Current time step.}
}
\value{
The updated \code{netsim_dat} main list object.
}
\description{
This function calls \code{summary} on each network being
simulated in \code{netsim}, provided \code{save.nwstats} and
\code{resimulate.network} are both \code{TRUE}. It records the
statistics represented by \code{nwstats.formula} in
\code{dat$stats$nwstats}, where \code{dat} is the main
\code{netsim_dat} class object.
}
\keyword{internal}
\keyword{netUtils}
