% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdx.R
\name{tedgelist_to_toggles}
\alias{tedgelist_to_toggles}
\title{Convert Timed Edgelist to Matrix of Toggles}
\usage{
tedgelist_to_toggles(tedgelist)
}
\arguments{
\item{tedgelist}{A timed edgelist, as produced by
\code{\link{as.data.frame.networkDynamic}}.}
}
\value{
The matrix of toggles corresponding to \code{tedgelist}.
}
\description{
Convert Timed Edgelist to Matrix of Toggles
}
\keyword{internal}
