% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_icm.bip}
\alias{test_icm.bip}
\title{Test the Model Output from a Stochastic Bipartite Individual Contact Model}
\usage{
test_icm.bip(x)
}
\arguments{
\item{x}{An object of class \code{icm.bip}.}
}
\description{
Tests whether the model output from an individual contact model
             is consistent with key balancing equations for compartment and
             flow sizes for each simulation for each time step.
}
\keyword{internal}
