% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test_icm}
\alias{test_icm}
\title{Test the Model Output from a Stochastic Individual Contact Model}
\usage{
test_icm(x)
}
\arguments{
\item{x}{an object of class \code{icm}.}
}
\description{
Tests whether the model output from an individual contact model
             is consistent with key balancing equations for compartment and
             flow sizes for each simulation for each time step.
}
\keyword{internal}

