% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test_net}
\alias{test_net}
\title{Test the Model Output from a Network Model}
\usage{
test_net(x)
}
\arguments{
\item{x}{an object of class \code{netsim}.}
}
\description{
Tests whether the model output from a network model is consistent
             with key balancing equations for compartment and flow sizes for
             each simulation for each time step.
}
\keyword{internal}

