% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdx.R
\name{toggles_to_diss_stats}
\alias{toggles_to_diss_stats}
\title{Convert Matrix of Toggles to Dissolution and Duration Statistics}
\usage{
toggles_to_diss_stats(toggles, coef.diss, nsteps, nw, time.start = 0L)
}
\arguments{
\item{toggles}{A matrix of toggles, as produced by
\code{\link[=tedgelist_to_toggles]{tedgelist_to_toggles()}}.}

\item{coef.diss}{Dissolution coefficients used in the simulation.}

\item{nsteps}{Number of time steps in the simulation.}

\item{nw}{Network used in the simulation.}

\item{time.start}{Starting time for the simulation.}
}
\value{
Named list containing dissolution and duration statistics matrices
and other related information.
}
\description{
Convert Matrix of Toggles to Dissolution and Duration Statistics
}
\keyword{internal}
