% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.end.horizon.R
\name{trigger_end_horizon}
\alias{trigger_end_horizon}
\title{Function to Trigger the End Horizon}
\usage{
trigger_end_horizon(dat)
}
\arguments{
\item{dat}{Main list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}
}
\value{
The updated \code{dat} main list object.
}
\description{
Function to Trigger the End Horizon
}
\details{
This function triggers the end horizon if a control \code{end.horizon} exists and
its \code{at} value is equal to the current timestep. The end horizon consists on
the removal of a set of modules from the module list.
}
\keyword{internal}
