% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.accessor.R
\name{uid-tools}
\alias{uid-tools}
\alias{get_uids}
\alias{get_indexes}
\title{Convert UIDs to Indexes and vice versa}
\usage{
get_uids(dat, indexes)

get_indexes(dat, uids)
}
\arguments{
\item{dat}{a Master list object of network models}

\item{indexes}{a vector of node indexes}

\item{uids}{a vector of node UIDs}
}
\value{
a vector of indexes or of UIDs
}
\description{
EpiModel refers to its nodes either by indexes, the position of
             the node in the \code{attr} vectors or by UIDs, unique
             identifiers allowing to refer to nodes even after they are
             deactivated
}
