% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{update_nwattr.bip}
\alias{update_nwattr.bip}
\title{Updates Vertex Attributes for Incoming Vertices}
\usage{
update_nwattr.bip(nw, newNodes, rules, curr.tab, t1.tab)
}
\arguments{
\item{nw}{The \code{networkDynamic.bip} object used in \code{netsim.bip} simulations.}

\item{newNodes}{Vector of nodal IDs for incoming nodes at the current time
step.}

\item{rules}{List of rules, one per attribute to be set, governing how to set
the values of each attribute.}

\item{curr.tab}{Current proportional distribution of all vertex attributes.}

\item{t1.tab}{Proportional distribution of all vertex attributes at the outset
of the simulation.}
}
\description{
Updates the vertex attributes on a network for new nodes incoming
             into that bipartite network, based on a set of rules for each attribute
             that the user specifies in \code{control.net.bip}.
}
\seealso{
\code{\link{copy_toall_attr.bip}}, \code{\link{get_attr_prop.bip}},
         \code{\link{update_nwattr.bip}}.
}
\keyword{internal}
\keyword{netUtils}
