% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.accessor.R
\name{update_uids}
\alias{update_uids}
\title{Create the uids for the new nodes}
\usage{
update_uids(dat, n.new)
}
\arguments{
\item{dat}{a Master list object of network models}

\item{n.new}{the number of new nodes to give \code{uid} to}
}
\value{
the Master list object of network models (\code{dat})
}
\description{
This function is called by `append_core_attr` and append new
uids to the created nodes. It also keeps track of the already used uids with
the /code{dat[["_last_uid"]]} variable
}
\keyword{internal}
