% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param2df.R
\name{use_scenario}
\alias{use_scenario}
\title{Apply a scenario object to a param.net object}
\usage{
use_scenario(param, scenario)
}
\description{
Apply a scenario object to a param.net object
}
\section{scenario}{

Can be made by make_scenarios_list from a scenarios.df
}

