% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.helpers.R
\name{validate_run}
\alias{validate_run}
\title{Ensures that the \code{run} sublist contains all the mandatory elements}
\usage{
validate_run(dat)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{netsim}}.}
}
\value{
The updated \code{netsim_dat} main list object.
}
\description{
Ensures that the \code{run} sublist contains all the mandatory elements
}
