% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose.R
\name{verbose.dcm.bip}
\alias{verbose.dcm.bip}
\title{Progress Print Module for Deterministic Bipartite Compartmental Models}
\usage{
verbose.dcm.bip(x, type, s = 1)
}
\arguments{
\item{x}{If the \code{type} is "startup", then an object of class
\code{control.icm.bip}, otherwise the all master data object in \code{icm.bip}
simulations.}

\item{type}{Progress type, either of "startup" for starting messages before
all simulations, or "progress" for time step specific messages.}

\item{s}{Current simulation number, if type is "progress".}
}
\description{
This function prints progress from deterministic bipartite compartmental
             models simulated with \code{dcm.bip} to the console.
}
\keyword{internal}
