% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose.R
\name{verbose.dcm}
\alias{verbose.dcm}
\title{Progress Print Module for Deterministic Compartmental Models}
\usage{
verbose.dcm(x, type, s = 1)
}
\arguments{
\item{x}{If the `type` is "startup", then an object of class
`control.icm`, otherwise the all master data object in `icm`
simulations.}

\item{type}{Progress type, either of "startup" for starting messages before
all simulations, or "progress" for time step specific messages.}

\item{s}{Current simulation number, if type is "progress".}
}
\description{
This function prints progress from deterministic compartmental
             models simulated with `dcm` to the console.
}
\keyword{internal}
