% Generated by roxygen2 (4.0.2): do not edit by hand
\name{verbose.icm}
\alias{verbose.icm}
\title{Progress Print Module for Stochastic Individual Contact Models}
\usage{
verbose.icm(x, type, s, at)
}
\arguments{
\item{x}{if the \code{type} is "startup", then an object of class
\code{control.icm}, otherwise the all master data object in \code{icm}
simulations.}

\item{type}{progress type, either of "startup" for starting messages before
all simulations, or "progress" for time step specific messages.}

\item{s}{current simulation number, if type is "progress".}

\item{at}{current time step, if type is "progress".}
}
\description{
This function prints progress from stochastic individual contact
             models simulated with \code{icm} to the console.
}
\keyword{internal}

