% Generated by roxygen2 (4.0.1): do not edit by hand
\name{verbose.net}
\alias{verbose.net}
\title{Progress Print Module for Stochastic Network Models}
\usage{
verbose.net(x, type, s, at)
}
\arguments{
\item{x}{if the \code{type} is "startup", then an object of class
\code{control.net}, otherwise the all master data object in \code{netsim}
simulations.}

\item{type}{progress type, either of "startup" for starting messages before
all simulations, or "progress" for time step specific messages.}

\item{s}{current simulation number, if type is "progress"}

\item{at}{current time step, if type is "progress"}
}
\description{
This function prints progress from stochastic network models
             simulated with \code{netsim} to the console.
}
\keyword{internal}

