% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.vital.R
\name{arrivals.2g.net}
\alias{arrivals.2g.net}
\title{Arrivals: netsim Module}
\usage{
arrivals.2g.net(dat, at)
}
\arguments{
\item{dat}{Master list object containing a `networkDynamic` object and other
initialization information passed from `\link{netsim}`.}

\item{at}{Current time step.}
}
\description{
This function simulates new arrivals into the network
  for use in `\link{netsim}` simulations.
}
\seealso{
`\link{netsim}`
}
\keyword{internal}
\keyword{netMod}
