% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.mod.vital.R
\name{arrivals.icm.bip}
\alias{arrivals.icm.bip}
\title{Arrivals: Bipartite icm Module}
\usage{
arrivals.icm.bip(dat, at)
}
\arguments{
\item{dat}{Main \code{icm_dat} class data object passed through \code{icm}
simulations.}

\item{at}{Current time step.}
}
\value{
The updated \code{icm_dat} class main data object.
}
\description{
This function simulates arrival for use in \code{\link[=icm]{icm()}}
simulations.
}
\seealso{
\code{\link[=icm]{icm()}}
}
\keyword{internal}
