% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.mod.vital.R
\name{arrivals.icm.bip}
\alias{arrivals.icm.bip}
\title{Arrivals: Bipartite icm Module}
\usage{
arrivals.icm.bip(dat, at)
}
\arguments{
\item{dat}{Main data object passed through \code{icm} simulations.}

\item{at}{Current time step.}
}
\value{
The updated \code{dat} main list object.
}
\description{
This function simulates arrival for use in \code{\link{icm}}
simulations.
}
\seealso{
\code{\link{icm}}
}
\keyword{internal}
