% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.vital.R
\name{arrivals.net.bip}
\alias{arrivals.net.bip}
\title{Arrivals: netsim.bip Module}
\usage{
arrivals.net.bip(dat, at)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic.bip} object and other
initialization information passed from \code{\link{netsim.bip}}.}

\item{at}{Current time step.}
}
\description{
This function simulates new arrivals into the network
  for use in \code{\link{netsim.bip}} simulations.
}
\seealso{
\code{\link{netsim.bip}}
}
\keyword{internal}
\keyword{netMod}
