% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{arrive_nodes}
\alias{arrive_nodes}
\title{Arrive New Nodes to the netsim_dat Object}
\usage{
arrive_nodes(dat, nArrivals)
}
\arguments{
\item{dat}{the \code{netsim_dat} object}

\item{nArrivals}{number of new nodes to arrive}
}
\value{
the updated \code{netsim_dat} object with \code{nArrivals} new nodes
added
}
\description{
Arrive New Nodes to the netsim_dat Object
}
\details{
\code{nArrivals} new nodes are added to the network data stored on
the \code{netsim_dat} object. If \code{tergmLite} is \code{FALSE}, these
nodes are activated from the current timestep onward. Attributes for the new
nodes must be set separately. On each static layer (see \code{\link[=netstatic]{netstatic()}}), the
new nodes are then placed under the layer's \code{arrivals} rule, which may set
the layer's grouping attribute for them and add their edges.

Note that this function only supports arriving new nodes; returning to an
active state nodes that were previously active in the network is not
supported.
}
