% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendo.R
\name{as.network.transmat}
\alias{as.network.transmat}
\title{Convert transmat Infection Tree into a network Object}
\usage{
\method{as.network}{transmat}(x, ...)
}
\arguments{
\item{x}{An object of class \code{transmat} to be converted into a
\code{network} class object.}

\item{...}{Unused.}
}
\value{
A \code{\link{network}} object.
}
\description{
Converts a transmission matrix from the \code{get_transmat}
function into a \link{network} class object.
}
\details{
When converting from a \code{transmat} to a \code{network} object, this
functions copies the edge attributes within the transmission matrix
(\code{'at'}, \code{'infDur'}, \code{'transProb'}, \code{'actRate'}, and
\code{'finalProb'}) into edge attributes on the network.
}
