% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendo.R
\name{as.network.transmat}
\alias{as.network.transmat}
\title{Convert transmat Infection Tree into a network Object}
\usage{
\method{as.network}{transmat}(x, ...)
}
\arguments{
\item{x}{An object of class \code{transmat} to be converted into a network
object.}

\item{...}{Unused.}
}
\value{
A \code{\link{network}} object.
}
\description{
Converts the edges of the infection tree described in the
             \code{\link{transmat}} object into a \code{\link{network}}
             object, copying in appropriate edge attributes for 'at',
             'infDur', 'transProb', 'actRate', and 'finalProb' and
             constructing a vertex attribute for 'at'.
}
