% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dendo.R
\name{as.network.transmat}
\alias{as.network.transmat}
\title{Converts transmat infection tree into a network object}
\usage{
\method{as.network}{transmat}(x, ...)
}
\arguments{
\item{x}{an object of class \code{transmat} to be converted into a network
object}

\item{...}{unused}
}
\description{
Converts the edges of the infection tree described in the
             transmat object into a \code{\link{network}} object, copying in
             appropriate edge attributes for 'at', 'infDur', 'transProb',
             'actRate', and 'finalProb' and constructing a vertex attribute
             for 'at'.
}

