% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reachable.R
\name{as_cumulative_edgelist}
\alias{as_cumulative_edgelist}
\title{Convert an object to a \code{cumulative_edgelist}}
\usage{
as_cumulative_edgelist(x)
}
\value{
A \code{cumulative_edgelist} object, a \code{data.frame} with at least the
following columns: \code{head}, \code{tail}, \code{start}, \code{stop}.
}
\description{
Convert an object to a \code{cumulative_edgelist}
}
\details{
The edges are active from time \code{start} to time \code{stop} included. If stop is
\code{NA}, the edge was not disolved in the simulation that generated the list.
}
