% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{auto_update_attr}
\alias{auto_update_attr}
\title{Updates Vertex Attributes for Incoming Vertices}
\usage{
auto_update_attr(dat, newNodes, curr.tab)
}
\arguments{
\item{dat}{Master object in `netsim` simulations.}

\item{newNodes}{Vector of nodal IDs for incoming nodes at the current time
step.}

\item{curr.tab}{Current proportional distribution of all vertex attributes.}
}
\description{
Updates the vertex attributes on a network for new nodes incoming
             into that network, based on a set of rules for each attribute
             that the user specifies in `control.net`.
}
\seealso{
`\link{copy_nwattr_to_datattr}`, `\link{get_attr_prop}`,
         `\link{auto_update_attr}`.
}
\keyword{internal}
\keyword{netUtils}
