% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{auto_update_attr}
\alias{auto_update_attr}
\title{Updates Vertex Attributes for Incoming Vertices}
\usage{
auto_update_attr(dat, newNodes, rules, curr.tab, t1.tab)
}
\arguments{
\item{dat}{Master object in \code{netsim} simulations.}

\item{newNodes}{Vector of nodal IDs for incoming nodes at the current time
step.}

\item{rules}{List of rules, one per attribute to be set, governing how to set
the values of each attribute.}

\item{curr.tab}{Current proportional distribution of all vertex attributes.}

\item{t1.tab}{Proportional distribution of all vertex attributes at the outset
of the simulation.}
}
\description{
Updates the vertex attributes on a network for new nodes incoming
             into that network, based on a set of rules for each attribute
             that the user specifies in \code{control.net}.
}
\seealso{
\code{\link{copy_nwattr_to_datattr}}, \code{\link{get_attr_prop}},
         \code{\link{auto_update_attr}}.
}
\keyword{internal}
\keyword{netUtils}
