% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{auto_update_attr}
\alias{auto_update_attr}
\title{Update Vertex Attributes for Incoming Vertices}
\usage{
auto_update_attr(dat, newNodes, curr.tab)
}
\arguments{
\item{dat}{Main data object passed through \code{netsim} simulations.}

\item{newNodes}{Vector of nodal IDs for incoming nodes at the current time
step.}

\item{curr.tab}{Current proportional distribution of all vertex attributes.}
}
\value{
The updated main \code{dat} object.
}
\description{
Updates the vertex attributes on a network for new nodes
             incoming into that network, based on a set of rules for each
             attribute that the user specifies in \code{\link{control.net}}.
}
\seealso{
\code{\link{copy_nwattr_to_datattr}}, \code{\link{get_attr_prop}},
         \code{\link{auto_update_attr}}.
}
\keyword{internal}
\keyword{netUtils}
