% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/net.utils.R
\name{bipvals}
\alias{bipvals}
\title{Vertex Attributes for Bipartite Network}
\usage{
bipvals(nw, mode, val)
}
\arguments{
\item{nw}{An object of class \code{network} or \code{networkDynamic}.}

\item{mode}{Mode number.}

\item{val}{Nodal attribute to return.}
}
\description{
Outputs static vertex attributes for a bipartite network for one
             specified mode.
}
\examples{
nw <- network.initialize(n = 10, bipartite = 5)
nw <- set.vertex.attribute(nw, "male", rep(0:1, each = 5))
bipvals(nw, mode = 1, "male")
}
\keyword{internal}
\keyword{netUtils}

