% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/icm.mod.vital.R
\name{births.icm}
\alias{births.icm}
\title{Births: icm Module}
\usage{
births.icm(dat, at)
}
\arguments{
\item{dat}{Master data list object.}

\item{at}{Current time step.}
}
\description{
This function simulates birth for use in \code{\link{icm}}
             simulations.
}
\seealso{
\code{\link{icm}}
}
\keyword{internal}

