% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.vital.R
\name{births.net}
\alias{births.net}
\title{Births: netsim Module}
\usage{
births.net(dat, at)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{at}{Current time step.}
}
\description{
This function simulates new births into the network
  for use in \code{\link{netsim}} simulations.
}
\seealso{
\code{\link{netsim}}
}
\keyword{internal}
\keyword{netMod}
