% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.accessor.R
\name{check_attr_lengths}
\alias{check_attr_lengths}
\title{Check that all \code{attr}ibutes in the master object are of equal
       length}
\usage{
check_attr_lengths(dat)
}
\arguments{
\item{dat}{a Master list object of network models}
}
\value{
invisible(TRUE) if everythin is correct. It throws an error otherwise
}
\description{
Check that all \code{attr}ibutes in the master object are of equal
       length
}
\keyword{internal}
\keyword{not_used}
