% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.accessor.R
\name{check_attr_lengths}
\alias{check_attr_lengths}
\title{Check that All Attributes in the Main Object are of Equal
Length}
\usage{
check_attr_lengths(dat)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{netsim}}.}
}
\value{
invisible(TRUE) if everything is correct; an error if not.
}
\description{
Check that All Attributes in the Main Object are of Equal
Length
}
\keyword{internal}
\keyword{not_used}
