% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.updater.R
\name{common_updater}
\alias{common_updater}
\title{Update Either the "param" or "control" List}
\usage{
common_updater(dat, type)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link[=netsim]{netsim()}}.}

\item{type}{Either \code{"param"} or \code{"control"}.}
}
\value{
The updated \code{netsim_dat} main list object.
}
\description{
Update Either the "param" or "control" List
}
\keyword{internal}
