% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{copy_datattr_to_nwattr}
\alias{copy_datattr_to_nwattr}
\title{Copy Vertex Attributes from the \code{dat} List to the Network Object}
\usage{
copy_datattr_to_nwattr(dat, network = seq_along(dat$nwparam))
}
\arguments{
\item{dat}{Main list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}
}
\value{
The updated \code{dat} main list object.
}
\description{
Copies the vertex attributes stored on the main \code{attr} list
             on \code{dat} to the network object on \code{dat}.
}
\seealso{
\code{\link{get_formula_term_attr}}, \code{\link{get_attr_prop}},
         \code{\link{auto_update_attr}}, and
         \code{\link{copy_nwattr_to_datattr}}.
}
\keyword{internal}
\keyword{netUtils}
