% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{copy_toall_attr}
\alias{copy_toall_attr}
\title{Copies Vertex Attributes in Formation Formula to attr List}
\usage{
copy_toall_attr(dat, at, nwterms)
}
\arguments{
\item{dat}{Master data object passed through \code{netsim} simulations.}

\item{at}{Current time step.}

\item{nwterms}{Vector of attributes on network object, usually as
output of \code{\link{get_formula_term_attr}}.}
}
\description{
Copies the vertex attributes stored on the network object to the
             master attr list in the dat data object.
}
\seealso{
\code{\link{get_formula_term_attr}}, \code{\link{get_attr_prop}},
         \code{\link{update_nwattr}}.
}
\keyword{internal}
\keyword{netUtils}
