% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.helpers.R
\name{create_dat_object}
\alias{create_dat_object}
\title{Create a Minimal netsim_dat Main List Object for a Network Model}
\usage{
create_dat_object(
  param = list(),
  init = list(),
  control = list(),
  run = list()
)
}
\arguments{
\item{param}{An \code{EpiModel} object of class \code{\link[=param.net]{param.net()}}.}

\item{init}{An \code{EpiModel} object of class \code{\link[=init.net]{init.net()}}.}

\item{control}{An \code{EpiModel} object of class \code{\link[=control.net]{control.net()}}.}

\item{run}{A \code{list} that will contains the objects created by
\code{\link[=netsim]{netsim()}} that are required for between step communication. This
list must be preserved for restarting models.}
}
\value{
A \code{netsim_dat} main list object.
}
\description{
This helper function populates a \code{netsim_dat} main list
object with the minimal required elements. All parameters are
optional. When none are given the resulting object is only a
shell list of class \code{netsim_dat} with the different named
elements defined as empty lists.
}
