% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm.inputs.R
\name{crosscheck.dcm}
\alias{crosscheck.dcm}
\title{Cross Checking of Inputs for Deterministic Compartmental Models}
\usage{
crosscheck.dcm(param, init, control)
}
\arguments{
\item{param}{An \code{EpiModel} object of class \code{\link[=param.dcm]{param.dcm()}}.}

\item{init}{An \code{EpiModel} object of class \code{\link[=init.dcm]{init.dcm()}}.}

\item{control}{An \code{EpiModel} object of class \code{\link[=control.dcm]{control.dcm()}}.}
}
\value{
This function returns no objects.
}
\description{
This function checks that the three parameter lists from
\code{\link[=param.dcm]{param.dcm()}}, \code{\link[=init.dcm]{init.dcm()}}, and
\code{\link[=control.dcm]{control.dcm()}} are consistent.
}
\keyword{internal}
