% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crosscheck.dcm}
\alias{crosscheck.dcm}
\title{Cross Checking of Inputs for Deterministic Compartmental Models}
\usage{
crosscheck.dcm(param, init, control)
}
\arguments{
\item{param}{an \code{EpiModel} object of class \code{\link{param.dcm}}.}

\item{init}{an \code{EpiModel} object of class \code{\link{init.dcm}}.}

\item{control}{an \code{EpiModel} object of class \code{\link{control.dcm}}.}
}
\value{
This function returns no objects.
}
\description{
This function checks that the three parameter lists from
             \code{\link{param.dcm}}, \code{\link{init.dcm}}, and
             \code{\link{control.dcm}} are consistent.
}
\keyword{internal}

