% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.inputs.R
\name{crosscheck.icm}
\alias{crosscheck.icm}
\title{Cross Checking of Inputs for Stochastic Individual Contact Models}
\usage{
crosscheck.icm(param, init, control)
}
\arguments{
\item{param}{An `EpiModel` object of class `\link{param.icm}`.}

\item{init}{An `EpiModel` object of class `\link{init.icm}`.}

\item{control}{An `EpiModel` object of class `\link{control.icm}`.}
}
\value{
This function returns no objects.
}
\description{
This function checks that the three parameter lists from
             `\link{param.icm}`, `\link{init.icm}`, and
             `\link{control.icm}` are consistent.
}
\keyword{internal}
