% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crosscheck.net}
\alias{crosscheck.net}
\title{Cross Checking of Inputs for Stochastic Network Models}
\usage{
crosscheck.net(x, param, init, control)
}
\arguments{
\item{x}{an \code{EpiModel} object of class \code{\link{netest}}.}

\item{param}{an \code{EpiModel} object of class \code{\link{param.net}}.}

\item{init}{an \code{EpiModel} object of class \code{\link{init.net}}.}

\item{control}{an \code{EpiModel} object of class \code{\link{control.net}}.}
}
\value{
This function returns no objects.
}
\description{
This function checks that the estimation object from
             \code{\link{netest}} and the three parameter lists from
             \code{\link{param.net}}, \code{\link{init.net}}, and
             \code{\link{control.net}} are consistent.
}
\keyword{internal}

