% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.inputs.R
\name{crosscheck.net}
\alias{crosscheck.net}
\title{Cross Checking of Inputs for Stochastic Network Models}
\usage{
crosscheck.net(x, param, init, control)
}
\arguments{
\item{x}{An \code{EpiModel} object of class \code{\link{netest}}.}

\item{param}{An \code{EpiModel} object of class \code{\link{param.net}}.}

\item{init}{An \code{EpiModel} object of class \code{\link{init.net}}.}

\item{control}{An \code{EpiModel} object of class \code{\link{control.net}}.}
}
\value{
This function returns no objects.
}
\description{
This function checks that the estimation object from
\code{\link{netest}} and the three parameter lists from
\code{\link{param.net}}, \code{\link{init.net}}, and
\code{\link{control.net}} are consistent.
}
\keyword{internal}
