% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/icm.mod.vital.R
\name{deaths.icm}
\alias{deaths.icm}
\title{Deaths: icm Module}
\usage{
deaths.icm(dat, at)
}
\arguments{
\item{dat}{Master data list object.}

\item{at}{Current time step.}
}
\description{
This function simulates death for use in \code{\link{icm}}
             simulations.
}
\seealso{
\code{\link{icm}}
}
\keyword{internal}

