% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deaths.net}
\alias{deaths.net}
\title{Deaths: netsim Module}
\usage{
deaths.net(all, at)
}
\arguments{
\item{all}{a list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{at}{current time step.}
}
\description{
This function simulates death for use in \link{netsim} simulations.
}
\seealso{
\code{\link{netsim}}
}
\keyword{internal}
\keyword{netMod}

