% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reachable.R
\name{dedup_cumulative_edgelist}
\alias{dedup_cumulative_edgelist}
\title{Deduplicate a cumulative edgelist by combining overlapping edges}
\usage{
dedup_cumulative_edgelist(el)
}
\arguments{
\item{el}{A cumulative edgelist with potentially overlapping edges}
}
\value{
A cumulative edgelist with no overlapping edges
}
\description{
Deduplicate a cumulative edgelist by combining overlapping edges
}
