% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete_attr}
\alias{delete_attr}
\title{Delete Elements from Attribute List}
\usage{
delete_attr(dat, ids)
}
\arguments{
\item{dat}{Master list object containing a sublist of attributes.}

\item{ids}{ID numbers to delete from the list.}
}
\description{
Deletes elements from the master attribute list.
}
\keyword{internal}
