% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deleteAttr}
\alias{deleteAttr}
\title{Delete Elements from Attribute List}
\usage{
deleteAttr(attrList, ids)
}
\arguments{
\item{attrList}{Attribute list.}

\item{ids}{ID numbers to delete from the list.}
}
\value{
The updated attribute list.
}
\description{
Deletes elements from the main attribute list.
}
\keyword{internal}
