% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.mod.vital.R
\name{departures.icm.bip}
\alias{departures.icm.bip}
\title{Departure: bipartite icm Module}
\usage{
departures.icm.bip(dat, at)
}
\arguments{
\item{dat}{Master data list object.}

\item{at}{Current time step.}
}
\description{
This function simulates departure for use in \code{\link{icm}}
             simulations.
}
\seealso{
\code{\link{icm}}
}
\keyword{internal}
