% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.vital.R
\name{departures.net}
\alias{departures.net}
\title{Departures: netsim Module}
\usage{
departures.net(dat, at)
}
\arguments{
\item{dat}{Main list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}

\item{at}{Current time step.}
}
\description{
This function simulates departure for use in \link{netsim}
       simulations.
}
\seealso{
\code{\link{netsim}}
}
\keyword{internal}
\keyword{netMod}
