% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{edgelist_meanage}
\alias{edgelist_meanage}
\title{Mean Age of Partnerships over Time}
\usage{
edgelist_meanage(el, diss_term = NULL, attribute = NULL)
}
\arguments{
\item{el}{If not passing \code{x}, a timed edgelist from a
\code{networkDynamic} object extracted with the
\code{as.data.frame.networkDynamic} function.}

\item{x}{An \code{EpiModel} object of class \code{\link{netest}}.}
}
\value{
A vector containing the mean edge age at each timestep.
}
\description{
Outputs a vector of mean ages of edges at a series of timesteps.
}
\details{
This function calculates the mean partnership age at each time step over
a dynamic network simulation from \code{\link{netest}}. These objects
contain the network, edgelist, and dissolution objects needed for the
calculation. Alternatively, one may pass in these objects separately if
\code{netest} was not used, or statistics were not requested after
the estimation.

Currently, the calculations are limited to those dissolution formulas with a
single homogenous dissolution (\code{~offset(edges)}). This functionality
will be expanded in future releases.
}
\examples{
# Initialize and parameterize the network model
nw <- network_initialize(n = 100)
formation <- ~edges
target.stats <- 50
coef.diss <- dissolution_coefs(dissolution = ~offset(edges), duration = 20)

# Model estimation
est <- netest(nw, formation, target.stats, coef.diss, verbose = FALSE)

# Simulate the network and extract a timed edgelist
dx <- netdx(est, nsims = 1, nsteps = 100, keep.tedgelist = TRUE,
      verbose = FALSE)
el <- as.data.frame(dx)

# Calculate ages directly from edgelist
mean_ages <- edgelist_meanage(el = el)
mean_ages

# Alternatively, netdx calculates these
dx$pages
identical(dx$pages[[1]], mean_ages)

}
\keyword{internal}
\keyword{netUtils}
