% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edges_correct}
\alias{edges_correct}
\title{Adjustment for the Edges Coefficient with Changing Network Size}
\usage{
edges_correct(dat, at)
}
\arguments{
\item{dat}{master object in \code{netsim} simulations.}

\item{at}{current time step.}
}
\description{
Adjusts the edges coefficient in a dynamic network model
             simulated in \code{\link{netsim}} to preserve the mean
             degree of nodes in the network.
}
\keyword{internal}

