% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.simnet.R
\name{edges_correct}
\alias{edges_correct}
\title{Adjustment for the Edges Coefficient with Changing Network Size}
\usage{
edges_correct(dat, at)
}
\arguments{
\item{dat}{Main list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}

\item{at}{Current time step.}
}
\value{
The updated \code{dat} main list object.
}
\description{
Adjusts the edges coefficient in a dynamic network model
simulated in \code{\link{netsim}} to preserve the mean
degree of nodes in the network. Requires \code{at >= 2}.
Maintains the \code{sim.num(.g2)} epi fields (initialized in
\code{\link{sim_nets_t1}}) for computing the coefficient
adjustment.
}
\keyword{internal}
