\docType{package}
\name{EpiModel-package}
\alias{EpiModel}
\alias{EpiModel-package}
\title{Mathematical Modeling of Infectious Disease}
\description{
  \tabular{ll}{ Package: \tab EpiModel\cr Type: \tab
  Package\cr Version: \tab 0.25\cr Date: \tab 2013-04-04\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr }
}
\details{
  The EpiModel package provides functions for building,
  solving, and plotting mathematical models of infectious
  disease.
}
\author{
  Samuel M. Jenness (\email{sjenness@uw.edu})
}
\references{
  \url{http://www.statnet.org}
}
\keyword{package}

