% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModel-package.r
\docType{package}
\name{EpiModel-package}
\alias{EpiModel-package}
\alias{EpiModel}
\title{Mathematical Modeling of Infectious Disease}
\description{
\tabular{ll}{
   Package: \tab EpiModel\cr
   Type: \tab Package\cr
   Version: \tab 1.5.0\cr
   Date: \tab 2017-06-01\cr
   License: \tab GPL-3\cr
   LazyLoad: \tab yes\cr
}
}
\details{
The EpiModel software package provides tools for building, solving, and
visualizing mathematical models of infectious disease. These tools allow users
to simulate epidemic models in multiple frameworks for both pedagogical
purposes ("base models") and novel research purposes ("extension models").
}
\references{
The EpiModel website is at \url{http://epimodel.org/}, and the source
            code is at \url{http://github.com/statnet/EpiModel}. Bug reports
            and feature requests may be filed there. We have also developed an
            extension package specifically for modeling HIV and related
            sexually transmitted infections, called EpiModelHIV and available
            on Github at \url{http://github.com/statnet/EpiModelHIV}
}
\keyword{package}
