% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_args}
\alias{get_args}
\title{Get Arguments from Parent Function}
\usage{
get_args(...)
}
\arguments{
\item{...}{Dot arguments in function.}
}
\description{
Gets the arguments and values from the parent function environment
             and returns them in a list.
}
\keyword{internal}

