% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_args}
\alias{get_args}
\title{Get Arguments from EpiModel Parameterization Functions}
\usage{
get_args(formal.args, dot.args)
}
\arguments{
\item{formal.args}{The output of \code{formals(sys.function())}.}

\item{dot.args}{The output of \code{list(...)}.}
}
\value{
A list of argument names and values.
}
\description{
Returns a list of argument names and values for use for
parameter processing functions.
}
\keyword{internal}
