% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{get_attr_prop}
\alias{get_attr_prop}
\title{Proportional Table of Vertex Attributes}
\usage{
get_attr_prop(dat, nwterms)
}
\arguments{
\item{nwterms}{Vector of attributes on network object, usually as
output of `\link{get_formula_term_attr}`.}

\item{nw}{The `networkDynamic` object contained in the `netsim`
simulation.}
}
\description{
Calculates the proportional distribution of each vertex attribute
             contained on the network.
}
\seealso{
`\link{get_formula_term_attr}`, `\link{copy_nwattr_to_datattr}`,
         `\link{auto_update_attr}`.
}
\keyword{internal}
\keyword{netUtils}
