% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{get_attr_prop}
\alias{get_attr_prop}
\title{Proportional Table of Vertex Attributes}
\usage{
get_attr_prop(dat, nwterms)
}
\arguments{
\item{nwterms}{Vector of attributes on network object, usually as
output of \code{\link{get_formula_term_attr}}.}

\item{nw}{The \code{networkDynamic} object contained in the \code{netsim}
simulation.}
}
\description{
Calculates the proportional distribution of each vertex
             attribute contained in the network.
}
\seealso{
\code{\link{get_formula_term_attr}},
         \code{\link{copy_nwattr_to_datattr}},
         \code{\link{auto_update_attr}}.
}
\keyword{internal}
\keyword{netUtils}
