% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{get_attr_prop}
\alias{get_attr_prop}
\title{Proportional Table of Vertex Attributes}
\usage{
get_attr_prop(dat, nwterms)
}
\arguments{
\item{dat}{Main list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}

\item{nwterms}{Vector of attributes on the network object, usually as
output of \code{\link{get_formula_term_attr}}.}
}
\value{
A table containing the proportional distribution of each attribute in
\code{nwterms}.
}
\description{
Calculates the proportional distribution of each vertex
attribute contained in a network.
}
\seealso{
\code{\link{get_formula_term_attr}},
\code{\link{copy_nwattr_to_datattr}},
\code{\link{auto_update_attr}}.
}
\keyword{internal}
\keyword{netUtils}
