% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{get_attr_prop}
\alias{get_attr_prop}
\title{Proportional Table of Vertex Attributes}
\usage{
get_attr_prop(dat, nwterms, only.formula = FALSE)
}
\arguments{
\item{nwterms}{Vector of attributes on network object, usually as
output of \code{\link{get_formula_term_attr}}.}

\item{only.formula}{Limit the tables to those terms only in formation model,
through \code{fterms}, otherwise output proportions for all attributes
on the network object.}

\item{nw}{The \code{networkDynamic} object contained in the \code{netsim}
simulation.}
}
\description{
Calculates the proportional distribution of each vertex attribute
             contained on the network, with a possible limitation to those
             attributes contained in the formation formula only.
}
\seealso{
\code{\link{get_formula_term_attr}}, \code{\link{copy_toall_attr}},
         \code{\link{update_nwattr}}.
}
\keyword{internal}
\keyword{netUtils}
