% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_cuml.R
\name{get_cumulative_degree}
\alias{get_cumulative_degree}
\title{Return the Cumulative Degree of a Set of Index Nodes}
\usage{
get_cumulative_degree(
  dat,
  index_posit_ids,
  networks = NULL,
  truncate = Inf,
  only.active.nodes = FALSE
)
}
\value{
A \code{data.frame} with 2 columns:
\itemize{
\item \code{index}: the unique ID (see \code{get_unique_ids}) of the
indexes.
\item \code{degree}: the cumulative degree of the index.
}
}
\description{
Return the Cumulative Degree of a Set of Index Nodes
}
\section{Cumulative Degree}{

The cumulative degree of a node is the number of edges connected to this
node at during the time window. The time window is by default all the steps
stored in the \code{cumulative_edgelist} or set by the \code{truncate} parameter.
}

