% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_cuml.R
\name{get_cumulative_edgelist}
\alias{get_cumulative_edgelist}
\title{Get a Cumulative Edgelist From the Specified Network}
\usage{
get_cumulative_edgelist(dat, network)
}
\arguments{
\item{dat}{a Master list object of network models}

\item{network}{the index of the network to extract the cumulative edgelist
from}
}
\value{
a cumulative edgelist in \code{data.frame} form with 4 columns:
         head and tail, the unique_ids (see \code{get_unique_ids}) of the
         nodes on the edge. Start and stop, the timestep where the edges
         started and stopped.
}
\description{
Get a Cumulative Edgelist From the Specified Network
}
