% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_cuml.R
\name{get_cumulative_edgelists_df}
\alias{get_cumulative_edgelists_df}
\title{Get the Cumulative Edgelists of a Model in \code{data.frame} Format}
\usage{
get_cumulative_edgelists_df(dat, networks = NULL)
}
\arguments{
\item{dat}{a Master list object of network models}

\item{networks}{the indexes of the networks to extract the cumulative
edgelists from. If NULL (default) extract all the cumulative
edgelists.}
}
\value{
a \code{data.frame} with 5 columns:
         head and tail, the unique_ids (see \code{get_unique_ids}) of the
         nodes on the edge. Start and stop, the time step where the edges
         started and stopped. network, the network on which the edge is.
}
\description{
Get the Cumulative Edgelists of a Model in \code{data.frame} Format
}
