% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.helpers.R
\name{get_current_timestep}
\alias{get_current_timestep}
\title{Return the Current Timestep}
\usage{
get_current_timestep(dat)
}
\arguments{
\item{dat}{Main list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}
}
\value{
The current timestep.
}
\description{
Return the Current Timestep
}
