% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_cuml.R
\name{get_edgelist}
\alias{get_edgelist}
\title{Get an Edgelist From the Specified Network}
\usage{
get_edgelist(dat, network)
}
\arguments{
\item{dat}{Main data object passed through \code{netsim} simulations.}

\item{network}{Numerical index of the network from which the edgelist should
be extracted. (May be > 1 for models with multiple overlapping
networks.)}
}
\value{
An edgelist in matrix form with two columns. Each column contains the
posit_ids (see \code{get_posit_ids}) of the nodes in each edge.
}
\description{
This function outputs an edgelist from the specified network,
             selecting the method depending on the stored network type.
}
