% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_cuml.R
\name{get_edgelist}
\alias{get_edgelist}
\title{Get an Edgelist From the Specified Network}
\usage{
get_edgelist(dat, at, network)
}
\arguments{
\item{dat}{a Master list object of network models}

\item{at}{the current timestep}

\item{network}{the index of the network to extract the edgelist from}
}
\value{
an edgelist in matrix form with the two columns. Each column
         contains the posit_ids (see \code{get_posit_ids})of the nodes in
         each edge
}
\description{
This function outputs an edgelist from the specified network.
             it chooses the right method depending on the type of network and
             the output is standardly formatted.
}
