% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{get_formula_term_attr}
\alias{get_formula_term_attr}
\title{Outputs ERGM Formula Attributes into a Character Vector}
\usage{
get_formula_term_attr(form, nw)
}
\arguments{
\item{form}{an ergm model formula}

\item{nw}{a network object}
}
\description{
Given a formation formula for a network model, outputs it into
             a character vector of vertex attributes to be used in \code{netsim}
             simulations.
}
