% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{get_network_term_attr}
\alias{get_network_term_attr}
\title{Outputs ERGM Formula Attributes into a Character Vector}
\usage{
get_network_term_attr(nw)
}
\arguments{
\item{nw}{a network object}
}
\description{
Given a simulated network, outputs into
             a character vector of vertex attributes to be used in \code{netsim}
             simulations.
}
