% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.fn.utils.R
\name{get_network_term_attr}
\alias{get_network_term_attr}
\title{Output Network Attributes into a Character Vector}
\usage{
get_network_term_attr(nw)
}
\arguments{
\item{nw}{A network object.}
}
\value{
A character vector of vertex attributes.
}
\description{
Given a simulated network, outputs a character vector of vertex
attributes to be used in \code{\link[=netsim]{netsim()}} simulations.
}
