% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_nwstats}
\alias{get_nwstats}
\title{Extract Network Statistics from Network Epidemic Model}
\usage{
get_nwstats(x, sim = 1)
}
\arguments{
\item{x}{an \code{EpiModel} object of class \code{\link{netsim}}.}

\item{sim}{simulation number of extracted network.}
}
\description{
Extracts a data frame of network statistics from a network
             epidemic model.
}
\examples{
\dontrun{
## Simulate SI epidemic on bipartite Bernoulli random graph
nw <- network.initialize(n = 100, bipartite = 50, directed = FALSE)
formation <- ~ edges
target.stats <- 50
dissolution <- ~ offset(edges)
duration <- 20
coef.diss <- dissolution_coefs(dissolution, duration)
est <- netest(nw,
               formation,
               dissolution,
               target.stats,
               coef.diss,
               verbose = FALSE)
param <- param.net(inf.prob = 0.3, inf.prob.m2 = 0.15)
init <- init.net(i.num = 10, i.num.m2 = 10)
control <- control.net(type = "SI", nsteps = 10, nsims = 3,
                       nwstats.formula = ~ edges + meandeg + degree(0:5),
                       verbose = FALSE)
mod <- netsim(est, param, init, control)

## Extract the network statistics from simulation 2
get_nwstats(mod, sim = 2)
}
}
\keyword{extract}

