% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_cuml.R
\name{get_partners}
\alias{get_partners}
\title{Get the Previous Partners of a Set of Indexes}
\usage{
get_partners(
  dat,
  index_posit_ids,
  networks = NULL,
  max.age = Inf,
  only.active = FALSE
)
}
\arguments{
\item{dat}{a Master list object of network models}

\item{index_posit_ids}{the positional IDs of the indexes of interest}

\item{networks}{the indexes of the networks to extract the partnerships from.
If NULL (default) extract from all networks.}

\item{max.age}{after how many steps a partnershipl that is no longer active
should be removed from the results (default = Inf)}

\item{only.active}{should the inactive partners be removed from the results
(default = FALSE)}
}
\value{
a \code{data.frame} with 5 columns:
         index and partner, the unique_ids (see \code{get_unique_ids}) of the
         indexes and partners respectively. Start and stop, the timestep
         where the edges started and stopped. network, the network on which
         the partnership is.
}
\description{
Get the Previous Partners of a Set of Indexes
}
