% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_sims}
\alias{get_sims}
\title{Extract Network Simulations}
\usage{
get_sims(x, sims)
}
\arguments{
\item{x}{An object of class \code{netsim}.}

\item{sims}{A vector of simulation numbers to retain in the output object.}
}
\description{
Subsets the entire \code{netsim} object to a subset of
             simulations, essentially functioning like a reverse of
             \code{merge}.
}
\keyword{extract}

