% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.R
\name{get_sims}
\alias{get_sims}
\title{Extract Network Simulations}
\usage{
get_sims(x, sims)
}
\arguments{
\item{x}{An object of class \code{netsim}.}

\item{sims}{A vector of simulation numbers to retain in the output object,
       or \code{"mean"} which automatically selects the one simulation with
       the number infected at the final time step closest to the mean across
       all simulations.}
}
\description{
Subsets the entire \code{netsim} object to a subset of
             simulations, essentially functioning like a reverse of
             \code{merge}.
}
\keyword{extract}

