% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{idgroup}
\alias{idgroup}
\title{Mode Numbers for Two-Group Network}
\usage{
idgroup(nw, ids)
}
\arguments{
\item{nw}{Object of class `network` or `networkDynamic`.}

\item{ids}{Vector of ID numbers for which the group number
should be returned.}
}
\description{
Outputs group numbers give ID numbers for a two-group network.
}
\examples{
nw <- network_initialize(n = 10)
nw <- set_vertex_attribute(nw, "group", rep(c(1,2), each = 5))
idgroup(nw)
idgroup(nw, ids = c(3, 6))

}
\keyword{internal}
\keyword{netUtils}
