% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.helpers.R
\name{increment_timestep}
\alias{increment_timestep}
\title{Increment the Current Timestep}
\usage{
increment_timestep(dat)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link[=netsim]{netsim()}}.}
}
\value{
The updated \code{netsim_dat} main list object.
}
\description{
This function adds 1 to the timestep counter stored in the
\code{netsim_dat} main list object.
}
\section{Mutability}{

This DOES NOT modify the \code{netsim_dat} object in place. The result must
be assigned back to \code{dat} in order to be registered:
\code{dat <- increment_timestep(dat)}.
}

