% Generated by roxygen2 (4.0.1): do not edit by hand
\name{infection.icm}
\alias{infection.icm}
\title{Primary Infection Module for icm}
\usage{
infection.icm(all, at)
}
\arguments{
\item{all}{master data list object.}

\item{at}{current time step.}
}
\description{
This function simulates the main infection process given the
             current state of the actors in the system.
}
\keyword{internal}

