% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.inputs.R
\name{init.icm}
\alias{init.icm}
\title{Initial Conditions for Stochastic Individual Contact Models}
\usage{
init.icm(s.num, i.num, r.num, s.num.g2, i.num.g2, r.num.g2, ...)
}
\arguments{
\item{s.num}{Number of initial susceptible persons. For two-group models,
this is the number of initial group 1 susceptible persons.}

\item{i.num}{Number of initial infected persons. For two-group models, this
is the number of initial group 1 infected persons.}

\item{r.num}{Number of initial recovered persons. For two-group models, this
is the number of initial group 1 recovered persons. This parameter is
only used for the \code{SIR} model type.}

\item{s.num.g2}{Number of initial susceptible persons in group 2. This
parameter is only used for two-group models.}

\item{i.num.g2}{Number of initial infected persons in group 2. This parameter
is only used for two-group models.}

\item{r.num.g2}{Number of initial recovered persons in group 2. This
parameter is only used for two-group \code{SIR} models.}

\item{...}{Additional initial conditions passed to model.}
}
\description{
Sets the initial conditions for stochastic individual contact
             models simulated with \code{icm}.
}
\details{
The initial conditions for a model solved with \code{\link{icm}} should be
input into the \code{init.icm} function. This function handles initial
conditions for both base models and original models using new modules. For
an overview of initial conditions for base ICM class models, consult the
\href{http://www.epimodel.org/tut.html}{Basic ICMs} tutorial.
}
\seealso{
Use \code{\link{param.icm}} to specify model parameters and
         \code{\link{control.icm}} to specify the control settings. Run the
         parameterized model with \code{\link{icm}}.
}
\keyword{parameterization}
