% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.init.R
\name{init_nets}
\alias{init_nets}
\title{Network Data and Stats Initialization}
\usage{
init_nets(dat, x)
}
\arguments{
\item{dat}{A main data object of class \code{netsim_dat} obtained from
\code{\link[=create_dat_object]{create_dat_object()}}, including the \code{control}
argument.}

\item{x}{Either a fitted network model object of class \code{netest}, or a
list of such objects.}
}
\value{
A \code{netsim_dat} class main data object with network data and
stats initialized.
}
\description{
This function initializes the network data and stats on the main
\code{netsim_dat} class data object.
}
\keyword{internal}
