% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.init.R
\name{init_pids}
\alias{init_pids}
\title{Persistent ID Initialization}
\usage{
init_pids(nw, groups, prefixes = c("F", "M"))
}
\arguments{
\item{nw}{An object of class \code{networkDynamic}.}

\item{prefixes}{Character string prefix for group-specific ID.}
}
\description{
This function initializes the persistent IDs for
             a \code{networkDynamic} object.
}
\details{
This function is used for \code{\link{netsim}} simulations
over two-group networks for populations with vital dynamics. Persistent IDs are
required in this situation because when new nodes are added to the
first group in a two-group network, the IDs for the second mode shift
upward. Persistent IDs allow for an analysis of disease transmission
chains for these simulations. These IDs are also invoked in the
\code{\link{arrivals.net}} module when the persistent IDs of incoming nodes
must be set.
}
\examples{
# Initialize network with 25 female and 75 male
nw <- network.initialize(100)
nw <- set.vertex.attribute(nw, "male", rep(c(25, 75), c(25,75)))

# Set persistent IDs using the default F/M prefix
nw <- init_pids(nw)
nw \%v\% "vertex.names"

# Use another prefix combination
nw <- init_pids(nw, c("A", "B"))
nw \%v\% "vertex.names"

}
\seealso{
\code{\link{initialize.pids}}
}
\keyword{internal}
\keyword{netMod}
