% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErgmTerms.R
\name{InitErgmTerm.absdiffnodemix}
\alias{InitErgmTerm.absdiffnodemix}
\title{Definition for absdiffnodemix ERGM Term}
\usage{
InitErgmTerm.absdiffnodemix(nw, arglist, ...)
}
\arguments{
\item{nw}{An object of class \code{network}.}

\item{arglist}{A list of arguments as specified in the \code{ergm.userterms}
package framework.}

\item{...}{Additional data passed into the function as specified in the
\code{ergm.userterms} package framework.}
}
\description{
This function defines and initialize the absdiffnodemix ERGM term
             that allows for targeting age homophily by race.
}
\details{
This ERGM user term was written to allow for age-based homophily in partnership
formation that is heterogeneous by race. The \code{absdiff} component allows
targets the distribution of age mixing on that continuous variable, and the
\code{nodemix} component differentiates this for black-black, black-white,
and white-white couples.
}
\author{
Steven M. Goodreau
}
