% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErgmTerms.R
\name{InitErgmTerm.fuzzynodematch}
\alias{InitErgmTerm.fuzzynodematch}
\alias{fuzzynodematch}
\title{Definition for fuzzynodematch ERGM Term}
\usage{
InitErgmTerm.fuzzynodematch(nw, arglist, ...)
}
\arguments{
\item{nw}{An object of class \code{network}.}

\item{arglist}{A list of arguments as specified in the \code{ergm.userterms}
package framework.}

\item{...}{Additional data passed into the function as specified in the
\code{ergm.userterms} package framework.}
}
\description{
This function defines and initializes the fuzzynodematch ERGM
             term that allows for vectorized homophily.
}
\details{
This ERGM user term was written to allow for vectorized homophily.  
The \code{attr} term argument should specify a character vertex attribute.
It is assumed that "venues" are encoded in the character string for a given
node as \code{vwyxz} where \code{v} is the literal character \code{"v"} and 
\code{w}, \code{x}, \code{y}, and \code{z} are integers between \code{0} and
\code{9}.  Multiple venues in the same character string should be separated 
by \code{"|"}.  A node with no venues may have any string not starting in 
\code{"v"}.  Thus, \code{"v0034"} indicates the single venue indexed by 
\code{34}, \code{"ego00089"} indicates no venue, and \code{"v0023|v1253"}
indicates the two venues indexed by \code{23} and \code{1253}.

If the \code{binary} term argument is \code{FALSE} (the default), the change
statistic for an on-toggle is the number of venues on which the two nodes 
match; if \code{binary} is \code{TRUE}, the change statistic for an 
on-toggle is \code{1} if the two nodes match on any venues, and \code{0} 
otherwise.
}
