% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.mod.init.R
\name{initialize.icm}
\alias{initialize.icm}
\title{Initialization: icm Module}
\usage{
initialize.icm(param, init, control)
}
\arguments{
\item{param}{An \code{EpiModel} object of class \code{\link{param.icm}}.}

\item{init}{An \code{EpiModel} object of class \code{\link{init.icm}}.}

\item{control}{An \code{EpiModel} object of class \code{\link{control.icm}}.}
}
\value{
The updated \code{netsim_dat} main list object.
}
\description{
This function initializes the main \code{dat} object on which
data are stored, and simulates disease status and other
attributes.
}
\keyword{internal}
