% Generated by roxygen2 (4.0.1): do not edit by hand
\name{initialize.icm}
\alias{initialize.icm}
\title{Initialization: icm Module}
\usage{
initialize.icm(param, init, control)
}
\arguments{
\item{param}{an \code{EpiModel} object of class \code{\link{param.icm}}.}

\item{init}{an \code{EpiModel} object of class \code{\link{init.icm}}.}

\item{control}{an \code{EpiModel} object of class \code{\link{control.icm}}.}
}
\description{
This function initializes the master \code{all} object on which
             data are stored, and simulates disease status and other attributes.
}
\keyword{internal}

