% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.init.R
\name{initialize.net}
\alias{initialize.net}
\title{Initialization: netsim Module}
\usage{
initialize.net(x, param, init, control, s)
}
\arguments{
\item{x}{An `EpiModel` object of class `\link{netest}`.}

\item{param}{An `EpiModel` object of class `\link{param.net}`.}

\item{init}{An `EpiModel` object of class `\link{init.net}`.}

\item{control}{An `EpiModel` object of class `\link{control.net}`.}

\item{s}{Simulation number, used for restarting dependent simulations.}
}
\description{
This function initializes the master `dat` object on which
             data are stored, simulates the initial state of the network, and
             simulates disease status and other attributes.
}
\keyword{internal}
