% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initialize.net}
\alias{initialize.net}
\title{Initialization: netsim Module}
\usage{
initialize.net(x, param, init, control, s)
}
\arguments{
\item{x}{an \code{EpiModel} object of class \code{\link{netest}}.}

\item{param}{an \code{EpiModel} object of class \code{\link{param.net}}.}

\item{init}{an \code{EpiModel} object of class \code{\link{init.net}}.}

\item{control}{an \code{EpiModel} object of class \code{\link{control.net}}.}

\item{s}{simulation number, used for restarting dependent simulations.}
}
\description{
This function initializes the master \code{dat} object on which
             data are stored, simulates the initial state of the network, and
             simulates disease status and other attributes.
}
\keyword{internal}

