% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.fn.accessor.R
\name{is_active_unique_ids}
\alias{is_active_unique_ids}
\title{Are These Nodes Active (Unique IDs)}
\usage{
is_active_unique_ids(dat, unique_ids)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link[=netsim]{netsim()}}.}

\item{unique_ids}{A vector of node unique identifiers.}
}
\value{
A logical vector with TRUE if the node is still active and FALSE
otherwise.
}
\description{
Are These Nodes Active (Unique IDs)
}
