% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdx.R
\name{make_dissolution_stats}
\alias{make_dissolution_stats}
\title{Calculate the Dissolution Statistics of a Network}
\usage{
make_dissolution_stats(sim.df, coef.diss, nsteps, verbose = TRUE)
}
\arguments{
\item{sim.df}{a list of network objects (one per simulation)}

\item{coef.diss}{the \code{coef.diss} element of \code{nwparam}}

\item{nsteps}{the number of simualted steps}

\item{verbose}{a verbosity toggle (default = TRUE)}
}
\value{
a \code{list} of dissolution statistics
}
\description{
Calculate the Dissolution Statistics of a Network
}
