% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdx.R
\name{make_formation_table}
\alias{make_formation_table}
\title{Calculate the Formation Statistics of a Network}
\usage{
make_formation_table(merged.stats, targets)
}
\arguments{
\item{merged.stats}{a matrix of \code{nsims * nsteps} rows, with a column for
each of the formation target}

\item{targets}{a \code{data.frame} of the formation targets with two columns:
"names" and "targets"}
}
\value{
a \code{data.frame} of the formation statistics
}
\description{
Calculate the Formation Statistics of a Network
}
