% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdx.R
\name{make_formation_table}
\alias{make_formation_table}
\title{Calculate the Formation Statistics of a Network}
\usage{
make_formation_table(stats, targets)
}
\arguments{
\item{stats}{A list of formation statistic matrices, one for each 
simulation, with one row for each simulated network if 
\code{dynamic == FALSE}, one row for each time step if 
\code{dynamic == TRUE}, and one column for each statistic.}

\item{targets}{A \code{data.frame} of the formation targets with two columns:
"names" and "targets".}
}
\value{
A \code{data.frame} of the formation statistics.
}
\description{
Calculate the Formation Statistics of a Network
}
\keyword{internal}
