% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mcat}
\alias{mcat}
\title{Write Out Test Progress to Console}
\usage{
mcat(test)
}
\arguments{
\item{test}{a character string with the name of a test.}
}
\description{
Writes the name of a test and \code{...} to console for
             showing testing progress.
}
\keyword{internal}

