% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{modeids}
\alias{modeids}
\title{ID Numbers for Bipartite Network}
\usage{
modeids(nw, mode)
}
\arguments{
\item{nw}{Object of class \code{network} or \code{networkDynamic}.}

\item{mode}{Mode number to return ID numbers for.}
}
\description{
Outputs ID numbers for a mode number for a bipartite network.
}
\examples{
nw <- network.initialize(10, bipartite = 5)
modeids(nw, mode = 2)

}
\seealso{
\code{\link{idmode}} provides the reverse functionality.
}
\keyword{internal}
\keyword{netUtils}

