% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.inputs.R
\name{multilayer}
\alias{multilayer}
\title{Specify Controls by Network}
\usage{
multilayer(...)
}
\arguments{
\item{...}{control arguments to apply to each network, with the index of the
network corresponding to the index of the control argument}
}
\value{
an object of class \code{multilayer} containing the specified
control arguments
}
\description{
This utility function allows specification of certain
\code{\link[=netsim]{netsim()}} controls to vary by network.  The
\code{\link[=netsim]{netsim()}} control arguments currently supporting
\code{multilayer} specifications are \code{nwstats.formula},
\code{set.control.ergm}, \code{set.control.tergm}, and
\code{tergmLite.track.duration}.
}
\seealso{
\code{\link[=netsim]{netsim()}} for passing a list of \code{\link[=netest]{netest()}} fits, one per layer.
The \href{https://epimodel.github.io/sismid/11_advanced/mod11-Tutorial.html}{Multi-Layer Networks}
chapter of the Network Modeling for Epidemics course materials works
through a two-layer model in full.
}
