% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.inputs.R
\name{multilayer}
\alias{multilayer}
\title{Specify Controls by Network}
\usage{
multilayer(...)
}
\arguments{
\item{...}{control arguments to apply to each network, with the index of the
network corresponding to the index of the control argument}
}
\value{
an object of class \code{multilayer} containing the specified
        control arguments
}
\description{
This utility function allows specification of certain
             \code{\link{netsim}} controls to vary by network.
}
