% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{netsim_cond_msg}
\alias{netsim_cond_msg}
\title{Message to Find in Which Module a \code{condition} Occurred}
\usage{
netsim_cond_msg(cond, module, at, msg)
}
\arguments{
\item{cond}{The type of \code{condition} handled (message, warning, error).}

\item{module}{The name of the module where the \code{condition} occurred.}

\item{at}{The time step the \code{condition} occurred.}

\item{msg}{The \code{condition}'s message.}
}
\value{
A formatted string describing where and when the \code{condition}
        occurred as well as the \code{condition}'s message.
}
\description{
This function returns a formatted string describing when, where,
             and why an error, message, or warning occurred.
}
\keyword{internal}
