% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/netsim.R
\name{netsim_parallel}
\alias{netsim_parallel}
\title{Stochastic Network Models in Parallel}
\usage{
netsim_parallel(x, param, init, control, merge = TRUE)
}
\arguments{
\item{x}{Fitted network model object, as an object of class \code{netest}.
Alternatively, if restarting a previous simulation, may be an object of
class \code{netsim}.}

\item{param}{Model parameters, as an object of class \code{param.net}.}

\item{init}{Initial conditions, as an object of class \code{init.net}.}

\item{control}{Control settings, as an object of class
       \code{control.net}.}

\item{merge}{If \code{TRUE}, merge parallel simulations into one \code{netsim}
       object after simulation.}
}
\description{
Simulates stochastic network epidemic models for infectious
             disease in parallel (deprecated).
}
\details{
This function is deprecated and has been replaced by the \code{netsim_par}
function in the \code{EpiModelHPC package}.
}
\seealso{
The EpiModelHPC repository may be found at \url{http://github.com/statnet/EpiModelHPC}.
}
\keyword{internal}

