% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_initialize}
\alias{network_initialize}
\title{Initialize Network Object}
\usage{
network_initialize(n)
}
\arguments{
\item{n}{Network size.}
}
\value{
Returns an object of class \code{network}.
}
\description{
Initialize an undirected network object for use in EpiModel
workflows.
}
\details{
This function is used in \code{EpiModel} workflows to initialize an empty
network object.  The network attributes \code{directed}, \code{bipartite},
\code{hyper}, \code{loops}, and \code{multiple} are set to \code{FALSE}.
}
\examples{
nw <- network_initialize(100)
nw

}
