% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.nwupdate.R
\name{nwupdate.net}
\alias{nwupdate.net}
\title{Dynamic Network Updates}
\usage{
nwupdate.net(dat, at)
}
\arguments{
\item{dat}{Main list object containing a full \code{networkDynamic} object
or networkLite edgelist (if using tergmLite), and other initialization
information passed from \code{\link{netsim}}.}

\item{at}{Current time step.}
}
\value{
The updated \code{dat} main list object.
}
\description{
This function handles all calls to the network object contained
             on the main \code{dat} object handled in \code{netsim}.
}
