% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.nwupdate.R
\name{nwupdate.net}
\alias{nwupdate.net}
\title{Dynamic Network Updates}
\usage{
nwupdate.net(dat, at)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link[=netsim]{netsim()}}.}

\item{at}{Current time step.}
}
\value{
The updated \code{netsim_dat} main list object.
}
\description{
This function handles all calls to the network object contained
on the main \code{netsim_dat} object handled in \code{netsim}.
}
