% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.nwupdate.R
\name{nwupdate.net}
\alias{nwupdate.net}
\title{EpiModel Network Writes}
\usage{
nwupdate.net(dat, at)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{at}{Current time step.}
}
\description{
This function handles all call to network object contained in
\code{dat$nw} during simulation, whether a direct or indirect manipulation.
}
