% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.nwupdate.R
\name{nwupdate.net}
\alias{nwupdate.net}
\title{Dynamic Network Updates}
\usage{
nwupdate.net(dat, at)
}
\arguments{
\item{dat}{Master list object containing a full `networkDynamic`` object
or networkLite edgelist (if using tergmLite), and other initialization
information passed from `\link{netsim}`.}

\item{at}{Current time step.}
}
\description{
This function handles all calls to the network object contained
             on the master dat object handled in `netsim`.
}
