% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{param.dcm}
\alias{param.dcm}
\title{Epidemic Parameters for Deterministic Compartmental Models}
\usage{
param.dcm(inf.prob, inter.eff, inter.start, act.rate, rec.rate, a.rate,
  ds.rate, di.rate, dr.rate, ...)
}
\arguments{
\item{inf.prob}{Probability of infection per transmissible act between
a susceptible and an infected person. In two-group models, this is the
probability of infection for the group 1 members.}

\item{inter.eff}{Efficacy of an intervention which affects the per-act
probability of infection. Efficacy is defined as 1 - the relative
hazard of infection given exposure to the intervention, compared to no
exposure.}

\item{inter.start}{Time step at which the intervention starts, between 1 and
the number of time steps specified in the model. This will default to
1 if the \code{inter.eff} is defined but this parameter is not.}

\item{act.rate}{Average number of transmissible acts per person per unit time.}

\item{rec.rate}{Average rate of recovery with immunity (in \code{SIR} models)
or re-susceptibility (in \code{SIS} models). The recovery rate is the
reciprocal of the disease duration. This parameter is only used for
\code{SIR} and \code{SIS} models.}

\item{a.rate}{Arrival or entry rate. For one-group models, the arrival rate is the
rate of new arrivals per person per unit time.}

\item{ds.rate}{Departure or exit rate for susceptible.}

\item{di.rate}{Departure or exit rate for infected.}

\item{dr.rate}{Departure or exit rate for recovered. This parameter is only used for
\code{SIR} models.}

\item{...}{Additional arguments passed to model.}
}
\description{
Sets the epidemic parameters for deterministic compartmental
             models simulated with \code{dcm}.
}
\details{
\code{param.dcm} sets the epidemic parameters for deterministic compartmental
models solved with the \code{\link{dcm}} function. The models may use the
base types, for which these parameters are used, or original model
specifications for which these parameters may be used (but not necessarily).
A detailed description of DCM parameterization for base models is found
in the
\href{http://statnet.github.io/tut/BasicDCMs.html}{Basic DCMs} tutorial.

For base models, the model specification will be selected as a function
of the model parameters entered here and the control settings in
\code{\link{control.dcm}}. One-group and two-group models are available, where
the former assumes a homogenous mixing in the population and the latter
assumes a purely heterogenous mixing between two distinct partitions in the
population (e.g., men and women). Specifying any group two parameters (those
with a \code{.g2}) implies the simulation of a two-group model. All the
parameters for a desired model type must be specified, even if they are zero.
}
\section{Sensitivity Analyses}{

\code{dcm} has been designed to easily run DCM sensitivity analyses, where a
series of models varying one or more of the model parameters is run. This is
possible by setting any parameter as a vector of length greater than one. See
both the example below and the
\href{http://statnet.github.io/tut/BasicDCMs.html}{Basic DCMs} tutorial.
}

\section{New Model Types}{

To build original model specifications outside of the base models, start
by consulting the \href{http://statnet.github.io/tut/NewDCMs.html}{Solving
New DCMs with EpiModel} tutorial. Briefly, an original model may use either
the existing model parameters named here, an original set of parameters, or
a combination of both. The \code{...} argument allows the user to pass an
arbitrary set of new model parameters into \code{param.dcm}. Whereas there are
strict checks for base models that the model parameters are valid,
parameter validity is the user's responsibility with these original models.
}

\seealso{
Use \code{\link{init.dcm}} to specify the initial conditions and
         \code{\link{control.dcm}} to specify the control settings. Run the
         parameterized model with \code{\link{dcm}}.
}
\keyword{parameterization}
