% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{param.icm}
\alias{param.icm}
\title{Epidemic Parameters for Stochastic Individual Contact Models}
\usage{
param.icm(inf.prob, inter.eff, inter.start, act.rate, rec.rate, a.rate,
  ds.rate, di.rate, dr.rate, ...)
}
\arguments{
\item{inf.prob}{Probability of infection per transmissible act between
a susceptible and an infected person. In two-group models, this is the
probability of infection for the group 1 members.}

\item{inter.eff}{Efficacy of an intervention which affects the per-act
probability of infection. Efficacy is defined as 1 - the relative
hazard of infection given exposure to the intervention, compared to no
exposure.}

\item{inter.start}{Time step at which the intervention starts, between 1 and
the number of time steps specified in the model. This will default to
1 if the \code{inter.eff} is defined but this parameter is not.}

\item{act.rate}{Average number of transmissible acts per person per unit time.}

\item{rec.rate}{Average rate of recovery with immunity (in \code{SIR} models)
or re-susceptibility (in \code{SIS} models). The recovery rate is the
reciprocal of the disease duration. This parameter is only used for
\code{SIR} and \code{SIS} models.}

\item{a.rate}{Arrival or entry rate. For one-group models, the arrival rate is the
rate of new arrivals per person per unit time.}

\item{ds.rate}{Departure or exit rate for susceptible.}

\item{di.rate}{Departure or exit rate for infected.}

\item{dr.rate}{Departure or exit rate for recovered. This parameter is only used for
\code{SIR} models.}

\item{...}{Additional arguments passed to model.}
}
\description{
Sets the epidemic parameters for stochastic individual contact
             models simulated with \code{icm}.
}
\details{
\code{param.icm} sets the epidemic parameters for the stochastic individual
contact models simulated with the \code{\link{icm}} function. Models
may use the base types, for which these parameters are used, or new process
modules which may use these parameters (but not necessarily). A detailed
description of ICM parameterization for base models is found in the
\href{http://statnet.github.io/tut/BasicICMs.html}{Basic ICMs} tutorial.

For base models, the model specification will be chosen as a result of
the model parameters entered here and the control settings in
\code{\link{control.icm}}. One-group and two-group models are available, where
the former assumes a homogenous mixing in the population and the latter
assumes a purely heterogenous mixing between two distinct partitions in the
population (e.g., men and women). Specifying any group two parameters (those
with a \code{.g2}) implies the simulation of a two-group model and requires use
of the appropriate function (.bip). All the parameters for a desired model
type must be specified, even if they are zero.
}
\section{New Modules}{

To build original models outside of the base models, new process modules
may be constructed to replace the existing modules or to supplement the existing
set. These are passed into the control settings in \code{\link{control.icm}}.
New modules may use either the existing model parameters named here, an
original set of parameters, or a combination of both. The \code{...} allows
the user to pass an arbitrary set of original model parameters into
\code{param.icm}. Whereas there are strict checks with default modules for
parameter validity, these checks are the user's responsibility with new modules.
}

\seealso{
Use \code{\link{init.icm}} to specify the initial conditions and
         \code{\link{control.icm}} to specify the control settings. Run the
         parameterized model with \code{\link{icm}}.
}
\keyword{parameterization}
