% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.dcm}
\alias{plot.dcm}
\title{Plot Data from a Deterministic Compartmental Epidemic Model}
\usage{
\method{plot}{dcm}(x, y, popfrac, run, col, lwd, lty, alpha, leg, leg.name,
  leg.cex, xlim, ylim, main, axs, add = FALSE, ...)
}
\arguments{
\item{x}{an \code{EpiModel} object of class \code{dcm}.}

\item{y}{output compartments or flows from \code{dcm} object to plot.}

\item{popfrac}{if \code{TRUE}, plot prevalence of values rather than numbers
(see details).}

\item{run}{run number to plot, for models with multiple runs (default is run 1).}

\item{col}{color for lines, either specified as a single color in a standard
R color format, or alternatively as a color palette from
\code{\link{RColorBrewer}} (see details).}

\item{lwd}{line width for output lines.}

\item{lty}{line type for output lines.}

\item{alpha}{transparency level for lines, where 0 = transparent and 1 = opaque
(see \code{\link{transco}}).}

\item{leg}{type of legend to plot. Values are "n" for no legend, "full" for
full legend, and "lim" for limited legend (see details).}

\item{leg.name}{character string to use for legend, with the default
determined automatically based on the \code{y} input.}

\item{leg.cex}{legend scale size (default of 0.8).}

\item{xlim}{x-axis scale limits for plot, with default based on model time
steps.}

\item{ylim}{y-axis scale limits for plot, with default based on range of data.}

\item{main}{character string for main plot title.}

\item{axs}{plot axis type (see \code{\link{par}} for details), with default
of "r".}

\item{add}{if \code{TRUE}, new plot window is not called and lines are added to
existing plot window.}

\item{...}{additional arguments to pass to main plot window (see
       \code{\link{plot.default}}).}
}
\description{
Plots epidemiological data from a deterministic compartment
             epidemic model solved with \code{dcm}.
}
\details{
This function plots epidemiological outcomes from a deterministic
compartmental model solved with \code{\link{dcm}}. Depending on the number of
model runs (sensitivity analyses) and number of groups, the default plot is
the fractional proportion of each compartment in the model over time. The
specific compartments or flows to plot may be set using the \code{y} parameter,
and in multiple run models the specific run may also be specified.
}
\section{The popfrac Argument}{

Compartment prevalences are the size of a compartment over some denominator.
To plot the raw numbers from any compartment, use \code{popfrac=FALSE}; this
is the default for any plots of flows. The \code{popfrac} parameter calculates
and plots the denominators of all specified compartments using these rules: 1)
for one-group models, the prevalence of any compartment is the compartment size
divided by the total population size; 2) for two-group models, the prevalence
of any compartment is the compartment size divided by the group size.
}

\section{Color Palettes}{

Since \code{\link{dcm}} supports multiple run sensitivity models, plotting
the results of such models uses a complex color scheme for distinguishing runs.
This is accomplished using the \code{\link{RColorBrewer}} color palettes, in
which includes a range of linked colors using named palettes. For
\code{plot.dcm}, one may either specify a brewer color palette listed in
\code{\link{brewer.pal.info}}, or alternatively a vector of standard R colors
(named, hexidecimal, or positive integers; see \code{\link{col2rgb}}).
}

\section{Plot Legends}{

There are three automatic legend types available, and the legend is
added by default for plots. To turn off the legend, use \code{leg="n"}. To
plot a legend with values for every line in a sensitivity analysis, use
\code{leg="full"}. With models with many runs, this may be visually
overwhelming. In those cases, use \code{leg="lim"} to plot a legend limited
to the highest and lowest of the varying parameter in the model. In cases
where the default legend names are not helpful, one may override those names
with the \code{leg.name} argument.
}
\examples{
# Deterministic SIR model with varying act rate
param <- param.dcm(trans.rate = 0.2, act.rate = 1:10,
                   rec.rate = 1/3, b.rate = 0.011, ds.rate = 0.01,
                   di.rate = 0.03, dr.rate = 0.01)
init <- init.dcm(s.num = 1000, i.num = 1, r.num = 0)
control <- control.dcm(type = "SIR", nsteps = 100, dt = 0.25)
mod <- dcm(param, init, control)

# Plot disease prevalence by default
plot(mod)

# Plot prevalence of susceptibles
plot(mod, y = "s.num", col = "Greys")

# Plot number of susceptibles
plot(mod, y = "s.num", popfrac = FALSE, col = "Greys")

# Plot multiple runs of multiple compartments together
plot(mod, y = c("s.num", "i.num"),
     run = 5, xlim = c(0, 50))
plot(mod, y = c("s.num", "i.num"),
     run = 10, lty = 2, leg = "n", add = TRUE)
}
\seealso{
\code{\link{dcm}}, \code{\link{brewer.pal.info}}
}
\keyword{plot}

