% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.netsim}
\alias{plot.netsim}
\title{Plot Data from a Stochastic Network Epidemic Model}
\usage{
\method{plot}{netsim}(x, type = "epi", y, popfrac = FALSE,
  sim.lines = FALSE, sims, sim.col, sim.lwd, sim.alpha, mean.line = TRUE,
  mean.smooth = TRUE, mean.col, mean.lwd = 2, mean.lty = 1, qnts = 0.5,
  qnts.col, qnts.alpha, qnts.smooth = TRUE, leg, leg.cex = 0.8, axs = "r",
  add = FALSE, network = 1, at = 1, col.status = FALSE,
  shp.bip = NULL, stats, targ.line = TRUE, targ.col, targ.lwd = 2,
  targ.lty = 2, plots.joined, ...)
}
\arguments{
\item{x}{An \code{EpiModel} model object of class \code{netsim}.}

\item{type}{Type of plot: \code{"epi"} for epidemic model results,
\code{"network"} for a static network plot (\code{plot.network}),
or \code{"formation"} for network formation statistics.}

\item{y}{Output compartments or flows from \code{icm} object to plot.}

\item{popfrac}{If \code{TRUE}, plot prevalence of values rather than numbers
(see details).}

\item{sim.lines}{If \code{TRUE}, plot individual simulation lines. Default is
to plot lines for one-group models but not for two-group models.}

\item{sims}{If \code{type="epi"} or \code{"formation"}, a vector of
simulation numbers to plot. If \code{type="network"}, a single
simulation number for network plot, or else \code{"min"} to plot the
simulation number with the lowest disease prevalence, \code{"max"} for
the simulation with the highest disease prevalence, or \code{"mean"}
for the simulation with the prevalance closest to the mean across
simulations at the specified time step.}

\item{sim.col}{Vector of any standard R color format for simulation lines.}

\item{sim.lwd}{Line width for simulation lines.}

\item{sim.alpha}{Transparency level for simulation lines, where 0 = transparent
and 1 = opaque (see \code{\link{transco}}).}

\item{mean.line}{If \code{TRUE}, plot mean of simulations across time.}

\item{mean.smooth}{If \code{TRUE}, use a lowess smoother on the mean line.}

\item{mean.col}{Vector of any standard R color format for mean lines.}

\item{mean.lwd}{Line width for mean lines.}

\item{mean.lty}{Line type for mean lines.}

\item{qnts}{If numeric, plot polygon of simulation quantiles based on the
range implied by the argument (see details). If \code{FALSE}, suppress
polygon from plot.}

\item{qnts.col}{Vector of any standard R color format for polygons.}

\item{qnts.alpha}{Transparency level for quantile polygons, where 0 =
transparent and 1 = opaque (see \code{\link{transco}}).}

\item{qnts.smooth}{If \code{TRUE}, use a lowess smoother on quantile polygons.}

\item{leg}{If \code{TRUE}, plot default legend.}

\item{leg.cex}{Legend scale size.}

\item{axs}{Plot axis type (see \code{\link{par}} for details), with default
to \code{"r"}.}

\item{add}{If \code{TRUE}, new plot window is not called and lines are added to
existing plot window.}

\item{network}{Network number, for simulations with multiple networks
representing the population.}

\item{at}{If \code{type="network"}, time step for network graph.}

\item{col.status}{If \code{TRUE} and \code{type="network"}, automatic disease
status colors (blue = susceptible, red = infected, green = recovered).}

\item{shp.bip}{If \code{type="network"} and a bipartite simulation, shapes
for the mode 2 vertices, with acceptable inputs of "triangle" and
"square". Mode 1 vertices will be circles.}

\item{stats}{If \code{type="formation"}, network statistics to plot, among
those specified in \code{nwstats.formula} of \code{\link{control.net}},
with the default to plot all statistics.}

\item{targ.line}{If \code{TRUE}, plot target or expected value line for
the statistic of interest.}

\item{targ.col}{Vector of standard R colors for target statistic lines, with
default colors based on \code{RColorBrewer} color palettes.}

\item{targ.lwd}{Line width for the line showing the target statistic values.}

\item{targ.lty}{Line type for the line showing the target statistic values.}

\item{plots.joined}{If \code{TRUE} and \code{type="formation"}, combine all
target statistics in one plot, versus one plot per target statistic if
\code{FALSE}.}

\item{...}{additional arguments to pass.}
}
\description{
Plots epidemiological and network data from a stochastic network
             model simulated with \code{netsim}.
}
\details{
This plot function can produce three types of plots with a stochastic network
model simulated through \code{\link{netsim}}:
\enumerate{
 \item \strong{\code{type="epi"}}: epidemic model results (e.g., disease
       prevalence and incidence) may be plotted.
 \item \strong{\code{type="network"}}: a static network plot will be generated.
       A static network plot of a dynamic network is a cross-sectional
       extraction of that dynamic network at a specific time point. This
       plotting function wraps the \code{\link{plot.network}} function in the
       \code{network} package. Consult the help page for \code{plot.network}
       for all of the plotting parameters. In addition, four plotting parameters
       specific to \code{netsim} plots are available: \code{sim}, \code{at},
       \code{col.status}, and \code{shp.bip}.
 \item \strong{\code{type="formation"}}: summary network statistics related to
       the network model formation are plotted. These plots are similar to the
       formation plots for \code{netdx} objects. When running a \code{netsim}
       simulation, one must specify there that \code{save.nwstats=TRUE}; the
       plot here will then show the network statistics requested explicitly in
       \code{nwstats.formula}, or will use the formation formula set in
       \code{netest} otherwise.
}

When \code{type="epi"}, this plotting function will extract the epidemiological
output from a model object of class \code{netsim} and plot the time series
data of disease prevalence and other results. The summary statistics that the
function calculates and plots are individual simulation lines, means of the
individual simulation lines, and quantiles of those individual simulation lines.
The mean line, toggled on with \code{mean.line=TRUE} is calculated as the row mean
across simulations at each time step.

Compartment prevalences are the size of a compartment over some denominator.
To plot the raw numbers from any compartment, use \code{popfrac=FALSE}; this
is the default for any plots of flows. The \code{popfrac} parameter calculates
and plots the denominators of all specified compartments using these rules: 1)
for one-group models, the prevalence of any compartment is the compartment size
divided by the total population size; 2) for two-group models, the prevalence
of any compartment is the compartment size divided by the group population size.
For any prevalences that are not automatically calculated, the
\code{\link{mutate_epi}} may be used to add new variables to the \code{netsim}
object to plot or analyze.

The quantiles show the range of outcome values within a certain specified
quantile range. By default, the interquartile range is shown: that is the
middle 50\% of the data. This is specified by \code{qnts=0.5}. To show the
middle 95\% of the data, specify \code{qnts=0.95}. To toggle off the polygons
where they are plotted by default, specify \code{qnts=FALSE}.

When \code{type="network"}, this function will plot cross sections of the simulated
networks at specified time steps. Because it is only possible to plot one time
step from one simulation at a time, it is necessary to enter these in the
\code{at} and \code{sims} parameters. To aide in visualizing representative
and extreme simulations at specific time steps, the sims parameter may be set
to \code{"mean"} to plot the simulation in which the disease prevalence is
closest to the average across all simulations, \code{"min"} to plot the simulation
in which the prevalence is lowest, and \code{"max"} to plot the simulation
in which the prevalence is highest.
}
\examples{
## Independent SI Model
# Initialize network and set network model parameters
nw <- network.initialize(n = 100, bipartite = 50, directed = FALSE)
formation <- ~edges
target.stats <- 50
coef.diss <- dissolution_coefs(dissolution = ~offset(edges), duration = 20)

# Estimate the network model
est <- netest(nw, formation, target.stats, coef.diss, verbose = FALSE)

# Simulate the epidemic model
param <- param.net(inf.prob = 0.3, inf.prob.m2 = 0.15)
init <- init.net(i.num = 10, i.num.m2 = 10)
control <- control.net(type = "SI", nsteps = 100, nsims = 5,
                       verbose = FALSE, save.nwstats = TRUE,
                       nwstats.formula = ~edges + meandeg + concurrent)
mod <- netsim(est, param, init, control)

# Plot epidemic trajectory (default type)
plot(mod, type = "epi")
plot(mod, type = "epi", popfrac = TRUE)
plot(mod, type = "epi", y = "si.flow", qnts = 1)

# Plot static networks
par(mar = c(0,0,0,0))
plot(mod, type = "network")

# Automatic coloring of infected nodes as red
par(mfrow = c(1, 2), mar = c(0, 0, 2, 0))
plot(mod, type = "network", main = "Min Prev | Time 50",
     col.status = TRUE, at = 50, sims = "min")
plot(mod, type = "network", main = "Max Prev | Time 50",
     col.status = TRUE, at = 50, sims = "max")

# Automatic shape by mode number (circle = mode 1)
par(mar = c(0,0,0,0))
plot(mod, type = "network", at = 50, col.status = TRUE, shp.bip = "square")
plot(mod, type = "network", at = 50, col.status = TRUE, shp.bip = "triangle")

# Plot formation statistics
par(mfrow = c(1,1), mar = c(3,3,1,1), mgp = c(2,1,0))
plot(mod, type = "formation")
plot(mod, type = "formation", plots.joined = FALSE)
plot(mod, type = "formation", sim = 2:4)
plot(mod, type = "formation", plots.joined = FALSE,
     stats = c("edges", "concurrent"))
plot(mod, type = "formation", stats = "meandeg",
     sim.lwd = 2, sim.col = "seagreen")

}
\seealso{
\code{\link{plot.network}} \code{\link{mutate_epi}}
}
\keyword{plot}

