% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dendo.R
\name{plot.transmat}
\alias{plot.transmat}
\title{plot transmat infection tree in one of several styles}
\usage{
\method{plot}{transmat}(tm, style = c("phylo", "network", "gv_tree",
  "diffusion"), ...)
}
\description{
plots the infection tree described in a transmat object in one of several styles: phylogentic tree, a network, a hierarchical tree, or a diffusion timeline.
}

