% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dendo.R
\name{plot.transmat}
\alias{plot.transmat}
\title{plot transmat infection tree in one of several styles}
\usage{
\method{plot}{transmat}(tm, style = c("cascade", "network", "gv_tree",
  "phylo"), ...)
}
\description{
plot transmat infection tree in one of several styles
}

