% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendo.R
\name{plot.transmat}
\alias{plot.transmat}
\title{Plot transmat Infection Tree in One of Several Styles}
\usage{
\method{plot}{transmat}(x, style = c("phylo", "network", "transmissionTimeline"), ...)
}
\arguments{
\item{x}{A \code{\link{transmat}} object to be plotted.}

\item{style}{Character name of plot style. One of "phylo", "network",
or "transmissionTimeline"}

\item{...}{Additional plot arguments to be passed to lower-level plot
functions (plot.network, plot.phylo, etc)}
}
\description{
Plots the infection tree described in a \code{\link{transmat}}
             object in one of two styles: a phylogenetic tree or an
             un-rooted network.
}
\details{
The phylo plot requires the \code{ape} package. The
\code{ndtv::transmissionTimeline} requires that the \code{ndtv} package
is installed. All of the options are essentially
}
\seealso{
\code{\link{plot.network}}, \code{\link[ape]{plot.phylo}}
}
