% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dendo.R
\name{plot.transmat}
\alias{plot.transmat}
\title{plot transmat infection tree in the style of a phylo tree}
\usage{
\method{plot}{transmat}(tm, ...)
}
\description{
plot transmat infection tree in the style of a phylo tree
}

