% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icm.mod.status.R
\name{recovery.icm.bip}
\alias{recovery.icm.bip}
\title{Recovery: icm Module}
\usage{
recovery.icm.bip(dat, at)
}
\arguments{
\item{dat}{Master data list object.}

\item{at}{Current time step.}
}
\description{
This function simulates recovery from the infected state
             either to an distinct recovered state (SIR model type) or back
             to a susceptible state (SIS model type), for use in
             \code{\link{icm}}.
}
\keyword{internal}
