% Generated by roxygen2 (4.0.1): do not edit by hand
\name{recovery.icm}
\alias{recovery.icm}
\title{Recovery: icm Module}
\usage{
recovery.icm(all, at)
}
\arguments{
\item{all}{master data list object.}

\item{at}{current time step.}
}
\description{
This function simulates recovery from the infected state
             either to an distinct recovered state (SIR model type) or back
             to a susceptible state (SIS model type), for use in
             \code{\link{icm}}.
}
\keyword{internal}

