% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resim_nets}
\alias{resim_nets}
\title{Resimulate Dynamic Network at Time 2+}
\usage{
resim_nets(all, at)
}
\arguments{
\item{at}{current time step.}

\item{x}{a master object passed through \code{\link{netsim}}.}
}
\description{
This function resimulates the dynamic network in stochastic
             network models simulated in \code{\link{netsim}} with dependence
             between the epidemic and demographic processes and the network
             structure.
}
\keyword{internal}
\keyword{netUtils}

