% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.simnet.R
\name{resim_nets}
\alias{resim_nets}
\title{Resimulate Dynamic Network at Time 2+}
\usage{
resim_nets(dat, at)
}
\arguments{
\item{dat}{Main data object passed through \code{netsim} simulations.}

\item{at}{Current time step.}
}
\description{
This function resimulates the dynamic network in stochastic
             network models simulated in \code{\link{netsim}} with dependence
             between the epidemic and demographic processes and the network
             structure.
}
\keyword{internal}
\keyword{netUtils}
