% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.simnet.R
\name{resim_nets}
\alias{resim_nets}
\title{Resimulate Dynamic Network at Time 2+}
\usage{
resim_nets(dat, at)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{netsim}}.}

\item{at}{Current time step.}
}
\value{
The updated \code{netsim_dat} main list object.
}
\description{
This function resimulates the dynamic network in stochastic
network models simulated in \code{\link{netsim}} with dependence
between the epidemic and demographic processes and the network
structure.
}
\keyword{internal}
\keyword{netUtils}
