% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveout.R
\name{saveout.dcm}
\alias{saveout.dcm}
\title{Save dcm Data to Output List Format}
\usage{
saveout.dcm(df, s, param, control, out = NULL)
}
\arguments{
\item{df}{Master object in `\link{dcm}` simulations.}

\item{s}{Current run number.}

\item{param}{Param list set in `\link{param.dcm}`}

\item{control}{Control list set in `\link{control.dcm}`.}

\item{out}{Out list passed back in for updating at runs 2+.}
}
\description{
This function transfers the data from the master `df`
             object to the output `out` object at the end of each
             simulation in `\link{dcm}`.
}
\keyword{internal}
