% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveout.R
\name{saveout.dcm}
\alias{saveout.dcm}
\title{Save dcm Data to Output List Format}
\usage{
saveout.dcm(df, s, param, control, out)
}
\arguments{
\item{df}{Master object in \code{\link{dcm}} simulations.}

\item{s}{Current run number.}

\item{param}{Param list set in \code{\link{param.dcm}}}

\item{control}{Control list set in \code{\link{control.dcm}}.}

\item{out}{Out list passed back in for updating at runs 2+.}
}
\description{
This function transfers the data from the master \code{df}
             object to the output \code{out} object at the end of each
             simulation in \code{\link{dcm}}.
}
\keyword{internal}

