% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveout.R
\name{saveout.icm}
\alias{saveout.icm}
\title{Save icm Data to Output List Format}
\usage{
saveout.icm(dat, s, out = NULL)
}
\arguments{
\item{dat}{Master object in `icm` simulations.}

\item{s}{Current run number.}

\item{out}{Out list passed back in for updating at runs 2+.}
}
\description{
This function transfers the data from the master `all`
             object to the output `out` object at the end of each
             simulation in `\link{icm}`.
}
\keyword{internal}
