% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveout.R
\name{saveout.icm}
\alias{saveout.icm}
\title{Save icm Data to Output List Format}
\usage{
saveout.icm(dat, s, out = NULL)
}
\arguments{
\item{dat}{Master object in \code{icm} simulations.}

\item{s}{Current run number.}

\item{out}{Out list passed back in for updating at runs 2+.}
}
\description{
This function transfers the data from the master \code{all}
             object to the output \code{out} object at the end of each
             simulation in \code{\link{icm}}.
}
\keyword{internal}

