% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveout.R
\name{saveout.net}
\alias{saveout.net}
\title{Save netsim Data to Output List Format}
\usage{
saveout.net(dat, s, out = NULL)
}
\arguments{
\item{dat}{Master object in `netsim` simulations.}

\item{s}{Current simulation number.}

\item{out}{Out list passed back in for updating at simulations 2+.}
}
\description{
This function transfers the data from the master `dat`
             object to the output `out` object at the end of each
             simulation in `\link{netsim}`.
}
\keyword{internal}
