% Generated by roxygen2 (4.0.2): do not edit by hand
\name{saveout.net}
\alias{saveout.net}
\title{Save netsim Data to Output List Format}
\usage{
saveout.net(all, s, out)
}
\arguments{
\item{all}{master object in \code{netsim} simulations.}

\item{s}{current simulation number.}

\item{out}{out list passed back in for updating at simulations 2+.}
}
\description{
This function transfers the data from the master \code{all}
             object to the output \code{out} object at the end of each
             simulation in \code{\link{netsim}}.
}
\keyword{internal}

