% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveout.R
\name{saveout.net}
\alias{saveout.net}
\title{Save netsim Data to Output List Format}
\usage{
saveout.net(dat, s, out = NULL)
}
\arguments{
\item{dat}{Main data object passed through \code{netsim} simulations.}

\item{s}{Current simulation number.}

\item{out}{Out list passed back in for updating at simulations 2+.}
}
\description{
This function transfers the data from the main \code{dat}
             object to the output \code{out} object at the end of each
             simulation in \code{\link{netsim}}.
}
\keyword{internal}
