% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.helpers.R
\name{set_current_timestep}
\alias{set_current_timestep}
\title{Set the Current Timestep}
\usage{
set_current_timestep(dat, timestep)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link[=netsim]{netsim()}}.}

\item{timestep}{The new value for the timestep.}
}
\value{
The updated \code{netsim_dat} main list object.
}
\description{
Changes the current timestep in the \code{netsim_dat} object.
Use with caution. This function exists to work around unforeseen
corner cases. In most situation, \code{increment_timestep} is
preferred.
}
\section{Mutability}{

This DOES NOT modify the \code{netsim_dat} object in place. The result must
be assigned back to \code{dat} in order to be registered:
\code{dat <- increment_timestep(dat)}.
}

