% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_nets}
\alias{sim_nets}
\title{Simulate Dynamic Network at Time 1}
\usage{
sim_nets(x, nw, nsteps, control)
}
\arguments{
\item{x}{an \code{EpiModel} object of class \code{\link{netest}}.}

\item{nw}{a \code{networkDynamic} object.}

\item{nsteps}{number of time steps to simulate the network over.}

\item{nsims}{number of independent network simulations.}

\item{control}{an \code{EpiModel} object of class \code{\link{control.net}}.}
}
\description{
This function simulates a dynamic network over one or multiple
             time steps, to be used in \code{\link{netsim}} models.
}
\keyword{internal}
\keyword{netUtils}

