% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/net.mod.simnet.R
\name{sim_nets}
\alias{sim_nets}
\title{Simulate Dynamic Network at Time 1}
\usage{
sim_nets(x, nw, nsteps, control)
}
\arguments{
\item{x}{An \code{EpiModel} object of class \code{\link{netest}}.}

\item{nw}{A \code{networkDynamic} object.}

\item{nsteps}{Number of time steps to simulate the network over.}

\item{control}{An \code{EpiModel} object of class \code{\link{control.net}}.}

\item{nsims}{Number of independent network simulations.}
}
\description{
This function simulates a dynamic network over one or multiple
             time steps, to be used in \code{\link{netsim}} models.
}
\keyword{internal}
\keyword{netUtils}

