% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.simnet.R
\name{sim_nets_t1}
\alias{sim_nets_t1}
\title{Initialize Networks Used in netsim}
\usage{
sim_nets_t1(dat)
}
\arguments{
\item{dat}{Main list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}
}
\value{
The updated \code{dat} main list object.
}
\description{
}
\keyword{internal}
\keyword{netUtils}
