% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.simnet.R
\name{sim_nets_t1}
\alias{sim_nets_t1}
\title{Initialize Networks Used in netsim}
\usage{
sim_nets_t1(dat)
}
\arguments{
\item{dat}{Main list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}
}
\value{
The updated \code{dat} main list object.
}
\description{
This function initializes the networks used in
\code{\link{netsim}}. The initial edge set for a given network
is obtained either from simulating the cross-sectional model
(if \code{edapprox == TRUE}) or from the \code{newnetwork}
element of the \code{netest} object (if
\code{edapprox == FALSE}). Once the initial edge sets are
determined, the first time step is simulated if
\code{resimulate.network == TRUE}, and all time steps are
simulated if \code{resimulate.network == FALSE}.
}
\keyword{internal}
\keyword{netUtils}
