% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.simnet.R
\name{sim_nets_t1}
\alias{sim_nets_t1}
\title{Simulate Initial Network at Time 1 for Model Initialization}
\usage{
sim_nets_t1(x, dat, nsteps)
}
\arguments{
\item{x}{An \code{EpiModel} object of class \code{\link{netest}}.}

\item{dat}{Main list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}

\item{nsteps}{For TERGMs, the number of time steps to simulate the network
over; for ERGMs, the number of independent network panels to simulate.}
}
\value{
The updated \code{dat} main list object.
}
\description{
This function simulates a dynamic network over one or multiple
             time steps for TERGMs or one or multiple cross-sectional network
             panels for ERGMs, for use in \code{\link{netsim}} modeling.
}
\keyword{internal}
\keyword{netUtils}
