% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.simnet.R
\name{sim_nets_t1}
\alias{sim_nets_t1}
\title{Simulate Initial Network at Time 1 for Model Initialization}
\usage{
sim_nets_t1(dat, network = 1L)
}
\arguments{
\item{dat}{Main list object containing a \code{networkDynamic} object and
other initialization information passed from \code{\link{netsim}}.}

\item{network}{Index of network to simulate.}
}
\value{
The updated \code{dat} main list object.
}
\description{
This function simulates a dynamic network over one or multiple
             time steps for TERGMs or one or multiple cross-sectional network
             panels for ERGMs, for use in \code{\link{netsim}} modeling.
}
\keyword{internal}
\keyword{netUtils}
