% Generated by roxygen2 (4.0.2): do not edit by hand
\name{split_bip}
\alias{split_bip}
\title{Update Attribute Values for a Bipartite Network}
\usage{
split_bip(all, var, val, nCurrM1, nCurrM2, nBirths, nBirthsM2)
}
\arguments{
\item{all}{master data object passed through \code{netsim} simulations.}

\item{var}{variable to update.}

\item{val}{fixed value to set for all incoming nodes.}

\item{nCurrM1}{number currently in mode 1.}

\item{nCurrM2}{number currently in mode 2.}

\item{nBirths}{number of births/entries in mode 1.}

\item{nBirthsM2}{number of births/entries in mode2.}
}
\description{
Adds new values for attributes in a bipartite network in which
             there may be births/entries in the first mode, which requires
             splitting the attribute vector into two, adding the new values,
             and re-concatenating the two updated vectors.
}
\keyword{internal}
\keyword{netUtils}

