% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.utils.R
\name{split_bip}
\alias{split_bip}
\title{Update Attribute Values for a Bipartite Network}
\usage{
split_bip(dat, var, val, nCurrM1, nCurrM2, nBirths, nBirthsM2)
}
\arguments{
\item{dat}{Master data object passed through \code{netsim} simulations.}

\item{var}{Variable to update.}

\item{val}{Fixed value to set for all incoming nodes.}

\item{nCurrM1}{Number currently in mode 1.}

\item{nCurrM2}{Number currently in mode 2.}

\item{nBirths}{Number of births/entries in mode 1.}

\item{nBirthsM2}{Number of births/entries in mode 2.}
}
\description{
Adds new values for attributes in a bipartite network in which
             there may be births/entries in the first mode, which requires
             splitting the attribute vector into two, adding the new values,
             and re-concatenating the two updated vectors.
}
\keyword{internal}
\keyword{netUtils}
