% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.netest}
\alias{summary.netest}
\title{Summary for Network Model Fit}
\usage{
\method{summary}{netest}(object, ...)
}
\arguments{
\item{object}{An \code{EpiModel} object of class \code{netest}.}

\item{...}{Additional summary function arguments.}
}
\description{
Prints the summary model fit statistics for an ERGM or STERGM fit.
}
\details{
This function is simply a wrapper function for \code{summary.ergm} and
\code{summary.stergm}. Additionally, if the edges dissolution approximation
was used to fit the temporal ERGM, then the dissolution coefficient information
will be printed.
}
\keyword{internal}

