% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_icm}
\alias{test_icm}
\title{Test the Model Output from a Stochastic Individual Contact Model}
\usage{
test_icm(x)
}
\arguments{
\item{x}{An object of class \code{icm}.}
}
\description{
Tests whether the model output from an individual contact model
is consistent with key balancing equations for compartment and
flow sizes for each simulation for each time step.
}
\keyword{internal}
