% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_netest}
\alias{trim_netest}
\title{Function to Reduce the Size of the \code{netest} Object}
\usage{
trim_netest(object)
}
\arguments{
\item{object}{A \code{netest} object.}
}
\value{
A \code{netest} object with formula environments and some networks
        removed.
}
\description{
Trims formula environments and some networks from the 
             \code{netest} object.
}
\details{
Removes \code{environment(object$constraints)},
         \code{environment(object$coef.diss$dissolution)},  
         \code{environment(object$formation)}, and
         \code{object$fit$newnetworks}.  When \code{edapprox = TRUE},
         also removes \code{object$fit$network} and
         \code{environment(object$fit$formula)}; when 
         \code{edapprox = FALSE}, also removes \code{object$fit$newnetwork} and
         all but \code{formation} and \code{dissolution} from 
         \code{environment(object$fit$formula)}.
         
         For the output to be usable in simulation, there should not be 
         substitutions in the formulas, other than \code{formation} and
         \code{dissolution} in \code{object$fit$formula} when 
         \code{edapprox = FALSE}.
}
