% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.accessor.R
\name{uid-tools}
\alias{uid-tools}
\alias{get_unique_ids}
\alias{get_posit_ids}
\title{Convert Unique identifiers from/to Positional identifiers}
\usage{
get_unique_ids(dat, posit_ids)

get_posit_ids(dat, unique_ids)
}
\arguments{
\item{dat}{a Master list object of network models}

\item{posit_ids}{a vector of node positional identifiers}

\item{unique_ids}{a vector of node unique identifiers}
}
\value{
a vector of unique or positional identifiers
}
\description{
EpiModel refers to its nodes either by positional identifiers
             (\code{posit_ids}), the position of the node in the \code{attr}
             vectors or by unique identifiers (\code{unique_ids}), allowing
             to refer to nodes even after they are deactivated
}
\section{Deactivated nodes}{

  When providing \code{unique_ids} of deactivated nodes to \code{get_posit_ids},
  \code{NA}s are returned instead and a warning is produced.
}

