% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.mod.updater.R
\name{update_list}
\alias{update_list}
\title{Update List \code{x} Using the Elements of List \code{new.x}.}
\usage{
update_list(x, new.x)
}
\arguments{
\item{x}{A list.}

\item{new.x}{A list.}
}
\value{
The full \code{x} list with the modifications added by \code{new.x}.
}
\description{
Update List \code{x} Using the Elements of List \code{new.x}.
}
\details{
This function updates list \code{x} by name. If \code{x} and \code{new.x}
elements are not named, the function will not work properly. If a function is
provided to replace an element that was originally not a function, this
function will be applied to the original value.
}
\keyword{internal}
