% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.accessor.R
\name{update_unique_ids}
\alias{update_unique_ids}
\title{Create the Unique Identifiers for New Nodes}
\usage{
update_unique_ids(dat, n.new)
}
\arguments{
\item{dat}{Main data object passed through \code{netsim} simulations.}

\item{n.new}{The number of new nodes to give unique_ids to.}
}
\value{
The main list object of network models (\code{dat}).
}
\description{
This function is called by \code{append_core_attr} and appends
new unique_ids to the created nodes. It also keeps track of the already used
unique_ids with the \code{dat[["_last_unique_id"]]} variable.
}
\keyword{internal}
