% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.accessor.R
\name{update_unique_ids}
\alias{update_unique_ids}
\title{Create the Unique Identifiers for New Nodes}
\usage{
update_unique_ids(dat, n.new)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{netsim}}.}

\item{n.new}{The number of new nodes to give \code{unique_ids} to.}
}
\value{
The updated \code{netsim_dat} main list object.
}
\description{
This function is called by \code{\link{append_core_attr}} and
appends new \code{unique_ids} to the created nodes. It also keeps track of
the already used \code{unique_ids} with the \code{dat[["_last_unique_id"]]}
variable.
}
\keyword{internal}
