% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.helpers.R
\name{validate_run}
\alias{validate_run}
\title{Ensures that the \code{run} sublist contains all the mandatory elements}
\usage{
validate_run(dat)
}
\description{
Ensures that the \code{run} sublist contains all the mandatory elements
}
