% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose.R
\name{verbose.dcm}
\alias{verbose.dcm}
\title{Progress Print Module for Deterministic Compartmental Models}
\usage{
verbose.dcm(x, type, s = 1)
}
\arguments{
\item{x}{If the \code{type} is "startup", then an object of class
\code{control.dcm}, otherwise the main \code{df} object in \code{dcm}
runs.}

\item{type}{Progress type, either of "startup" for starting messages before
all runs, or "progress" for time step specific messages.}

\item{s}{Current run number, if type is "progress".}
}
\description{
This function prints progress from deterministic compartmental
models simulated with \code{\link[=dcm]{dcm()}} to the console.
}
\keyword{internal}
