% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/verbose.R
\name{verbose.dcm}
\alias{verbose.dcm}
\title{Progress Print Module for Deterministic Compartmental Models}
\usage{
verbose.dcm(x, type, s)
}
\arguments{
\item{x}{If the \code{type} is "startup", then an object of class
\code{control.icm}, otherwise the all master data object in \code{icm}
simulations.}

\item{type}{Progress type, either of "startup" for starting messages before
all simulations, or "progress" for time step specific messages.}

\item{s}{Current simulation number, if type is "progress".}
}
\description{
This function prints progress from deterministic compartmental
             models simulated with \code{dcm} to the console.
}
\keyword{internal}

