% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose.R
\name{verbose.net}
\alias{verbose.net}
\title{Progress Print Module for Stochastic Network Models}
\usage{
verbose.net(x, type, s = 1, at = 2)
}
\arguments{
\item{x}{If the \code{type} is "startup", then an object of class
\code{control.net}, otherwise the all master data object in \code{netsim}
simulations.}

\item{type}{Progress type, either of "startup" for starting messages before
all simulations, or "progress" for time step specific messages.}

\item{s}{Current simulation number, if type is "progress"}

\item{at}{Current time step, if type is "progress"}
}
\description{
This function prints progress from stochastic network models
             simulated with \code{netsim} to the console.
}
\keyword{internal}
